% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_coordinate.R
\name{absolute_coordinate}
\alias{absolute_coordinate}
\title{absolute_coordinate}
\usage{
absolute_coordinate(coord_string, ...)
}
\arguments{
\item{coord_string}{coord_string}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An string.
}
\description{
Convert a coordinate to an absolute coordinate string (B12 -> $B$12)

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\examples{
\dontrun{
absolute_coordinate("A1")
}

}
