% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsoluteAnchor.R
\name{AbsoluteAnchor}
\alias{AbsoluteAnchor}
\title{AbsoluteAnchor}
\usage{
AbsoluteAnchor(pos = NULL, ext = NULL, ...)
}
\arguments{
\item{pos}{pos}

\item{ext}{ext}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
AbsoluteAnchor()
}

}
