% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table.R
\name{add_table}
\alias{add_table}
\title{Check for duplicate name in definedNames and other worksheet tables before adding table.}
\usage{
add_table(py_obj, table)
}
\arguments{
\item{py_obj}{py_obj}

\item{table}{table}
}
\value{
An openpyxl Python object.
}
\description{
Check for duplicate name in definedNames and other worksheet tables before adding table.
}
\examples{
\dontrun{
wb <- Workbook()
ws <- wb$active
tab = Table(displayName="Table1", ref="A1:E5")
ws |> add_table(tab)
ws$tables
}

}
