% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjPoint2D.R
\name{AdjPoint2D}
\alias{AdjPoint2D}
\title{AdjPoint2D}
\usage{
AdjPoint2D(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
AdjPoint2D(x = 1L, y = 2L)
}

}
