% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustHandleList.R
\name{AdjustHandleList}
\alias{AdjustHandleList}
\title{AdjustHandleList}
\usage{
AdjustHandleList(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
AdjustHandleList()
}

}
