% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{Alignment}
\alias{Alignment}
\title{Alignment}
\usage{
Alignment(
  horizontal = NULL,
  vertical = NULL,
  textRotation = 0L,
  wrapText = NULL,
  shrinkToFit = NULL,
  indent = 0L,
  relativeIndent = 0L,
  justifyLastLine = NULL,
  readingOrder = 0L,
  text_rotation = NULL,
  wrap_text = NULL,
  shrink_to_fit = NULL,
  mergeCell = NULL,
  ...
)
}
\arguments{
\item{horizontal}{horizontal}

\item{vertical}{vertical}

\item{textRotation}{textRotation}

\item{wrapText}{wrapText}

\item{shrinkToFit}{shrinkToFit}

\item{indent}{indent}

\item{relativeIndent}{relativeIndent}

\item{justifyLastLine}{justifyLastLine}

\item{readingOrder}{readingOrder}

\item{text_rotation}{text_rotation}

\item{wrap_text}{wrap_text}

\item{shrink_to_fit}{shrink_to_fit}

\item{mergeCell}{mergeCell}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Alignment options for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.alignment.html}
}
\examples{
\dontrun{
Alignment()
}

}
