% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaCeilingEffect.R
\name{AlphaCeilingEffect}
\alias{AlphaCeilingEffect}
\title{AlphaCeilingEffect}
\usage{
AlphaCeilingEffect(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
AlphaCeilingEffect()
}

}
