% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaFloorEffect.R
\name{AlphaFloorEffect}
\alias{AlphaFloorEffect}
\title{AlphaFloorEffect}
\usage{
AlphaFloorEffect(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
AlphaFloorEffect()
}

}
