% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaModulateEffect.R
\name{AlphaModulateEffect}
\alias{AlphaModulateEffect}
\title{AlphaModulateEffect}
\usage{
AlphaModulateEffect(cont = NULL, ...)
}
\arguments{
\item{cont}{cont}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
AlphaModulateEffect(cont = EffectContainer(type = "tree"))
}

}
