% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaModulateFixedEffect.R
\name{AlphaModulateFixedEffect}
\alias{AlphaModulateFixedEffect}
\title{AlphaModulateFixedEffect}
\usage{
AlphaModulateFixedEffect(amt = NULL, ...)
}
\arguments{
\item{amt}{amt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
AlphaModulateFixedEffect(amt = 1L)
}

}
