% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaReplaceEffect.R
\name{AlphaReplaceEffect}
\alias{AlphaReplaceEffect}
\title{AlphaReplaceEffect}
\usage{
AlphaReplaceEffect(a = NULL, ...)
}
\arguments{
\item{a}{a}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
AlphaReplaceEffect(a = 1L)
}

}
