% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnchorClientData.R
\name{AnchorClientData}
\alias{AnchorClientData}
\title{AnchorClientData}
\usage{
AnchorClientData(fLocksWithSheet = NULL, fPrintsWithSheet = NULL, ...)
}
\arguments{
\item{fLocksWithSheet}{fLocksWithSheet}

\item{fPrintsWithSheet}{fPrintsWithSheet}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
AnchorClientData()
}

}
