% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnchorMarker.R
\name{AnchorMarker}
\alias{AnchorMarker}
\title{AnchorMarker}
\usage{
AnchorMarker(col = 0L, colOff = 0L, row = 0L, rowOff = 0L, ...)
}
\arguments{
\item{col}{col}

\item{colOff}{colOff}

\item{row}{row}

\item{rowOff}{rowOff}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
AnchorMarker()
}

}
