% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_to_degrees.R
\name{angle_to_degrees}
\alias{angle_to_degrees}
\title{angle_to_degrees}
\usage{
angle_to_degrees(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
angle_to_degrees(value = 1000L)
}

}
