% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaChart.R
\name{AreaChart}
\alias{AreaChart}
\title{AreaChart}
\usage{
AreaChart(axId = NULL, extLst = NULL, ...)
}
\arguments{
\item{axId}{axId}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.area_chart.html}
}
\examples{
\dontrun{
AreaChart()
}

}
