% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaChart3D.R
\name{AreaChart3D}
\alias{AreaChart3D}
\title{AreaChart3D}
\usage{
AreaChart3D(gapDepth = NULL, ...)
}
\arguments{
\item{gapDepth}{gapDepth}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.area_chart.html}
}
\examples{
\dontrun{
AreaChart3D()
}

}
