% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayDescriptor.R
\name{ArrayDescriptor}
\alias{ArrayDescriptor}
\title{ArrayDescriptor}
\usage{
ArrayDescriptor(key = NULL, ...)
}
\arguments{
\item{key}{key}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.cell_style.html}
}
\examples{
\dontrun{
ArrayDescriptor()
}

}
