% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayFormula.R
\name{ArrayFormula}
\alias{ArrayFormula}
\title{ArrayFormula}
\usage{
ArrayFormula(ref, text = NULL, ...)
}
\arguments{
\item{ref}{ref}

\item{text}{text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.formula.html}
}
\examples{
\dontrun{
ArrayFormula(ref = "A1")
}

}
