% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthorList.R
\name{AuthorList}
\alias{AuthorList}
\title{AuthorList}
\usage{
AuthorList(author = list(), ...)
}
\arguments{
\item{author}{author}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.comments.author.html}
}
\examples{
\dontrun{
AuthorList(author = c("author1", "author2"))
}

}
