% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFilter.R
\name{AutoFilter}
\alias{AutoFilter}
\title{AutoFilter}
\usage{
AutoFilter(
  ref = NULL,
  filterColumn = list(),
  sortState = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{ref}{ref}

\item{filterColumn}{filterColumn}

\item{sortState}{sortState}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
AutoFilter()
}

}
