% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutonumberBullet.R
\name{AutonumberBullet}
\alias{AutonumberBullet}
\title{AutonumberBullet}
\usage{
AutonumberBullet(type = NULL, startAt = NULL, ...)
}
\arguments{
\item{type}{type}

\item{startAt}{startAt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
AutonumberBullet(type = "hindiNumPeriod", startAt = 1L)
}

}
