% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoid_duplicate_name.R
\name{avoid_duplicate_name}
\alias{avoid_duplicate_name}
\title{avoid_duplicate_name}
\usage{
avoid_duplicate_name(names, value, ...)
}
\arguments{
\item{names}{names}

\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An character string.
}
\description{
Naive check to see whether name already exists.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.child.html}
}
\details{
If name does exist suggest a name using an incrementer
Duplicates are case insensitive
}
\examples{
\dontrun{
avoid_duplicate_name(names = "name", value = "name")
}

}
