% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxDataSource.R
\name{AxDataSource}
\alias{AxDataSource}
\title{AxDataSource}
\usage{
AxDataSource(
  numRef = NULL,
  numLit = NULL,
  strRef = NULL,
  strLit = NULL,
  multiLvlStrRef = NULL,
  ...
)
}
\arguments{
\item{numRef}{numRef}

\item{numLit}{numLit}

\item{strRef}{strRef}

\item{strLit}{strLit}

\item{multiLvlStrRef}{multiLvlStrRef}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
AxDataSource(numRef = NumRef())
}

}
