% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backdrop.R
\name{Backdrop}
\alias{Backdrop}
\title{Backdrop}
\usage{
Backdrop(anchor = NULL, norm = NULL, up = NULL, extLst = NULL, ...)
}
\arguments{
\item{anchor}{anchor}

\item{norm}{norm}

\item{up}{up}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Backdrop(
  anchor = Point3D(x = 1L, y = 2L, z = 3L),
  norm = Vector3D(dx = 1L, dy = 2L, dz = 3L),
  up = Vector3D(dx = 1L, dy = 2L, dz = 3L)
)
}

}
