% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BandFormat.R
\name{BandFormat}
\alias{BandFormat}
\title{BandFormat}
\usage{
BandFormat(idx = 0L, spPr = NULL, ...)
}
\arguments{
\item{idx}{idx}

\item{spPr}{spPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.surface_chart.html}
}
\examples{
\dontrun{
BandFormat()
}

}
