% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BandFormatList.R
\name{BandFormatList}
\alias{BandFormatList}
\title{BandFormatList}
\usage{
BandFormatList(bandFmt = list(), ...)
}
\arguments{
\item{bandFmt}{bandFmt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.surface_chart.html}
}
\examples{
\dontrun{
BandFormatList()
}

}
