% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarChart.R
\name{BarChart}
\alias{BarChart}
\title{BarChart}
\usage{
BarChart(gapWidth = 150L, overlap = NULL, serLines = NULL, extLst = NULL, ...)
}
\arguments{
\item{gapWidth}{gapWidth}

\item{overlap}{overlap}

\item{serLines}{serLines}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.bar_chart.html}
}
\examples{
\dontrun{
BarChart()
}

}
