% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarChart3D.R
\name{BarChart3D}
\alias{BarChart3D}
\title{BarChart3D}
\usage{
BarChart3D(
  gapWidth = 150L,
  gapDepth = 150L,
  shape = NULL,
  serLines = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{gapWidth}{gapWidth}

\item{gapDepth}{gapDepth}

\item{shape}{shape}

\item{serLines}{serLines}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.bar_chart.html}
}
\examples{
\dontrun{
BarChart3D()
}

}
