% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bevel.R
\name{Bevel}
\alias{Bevel}
\title{Bevel}
\usage{
Bevel(w = NULL, h = NULL, prst = NULL, ...)
}
\arguments{
\item{w}{w integer}

\item{h}{h integer}

\item{prst}{prst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Bevel(w = 1L, h = 1L)
}

}
