% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiLevelEffect.R
\name{BiLevelEffect}
\alias{BiLevelEffect}
\title{BiLevelEffect}
\usage{
BiLevelEffect(thresh = NULL, ...)
}
\arguments{
\item{thresh}{thresh}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
BiLevelEffect(thresh = 1L)
}

}
