% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blip.R
\name{Blip}
\alias{Blip}
\title{Blip}
\usage{
Blip(
  cstate = NULL,
  embed = NULL,
  link = NULL,
  noGrp = NULL,
  noSelect = NULL,
  noRot = NULL,
  noChangeAspect = NULL,
  noMove = NULL,
  noResize = NULL,
  noEditPoints = NULL,
  noAdjustHandles = NULL,
  noChangeArrowheads = NULL,
  noChangeShapeType = NULL,
  extLst = NULL,
  alphaBiLevel = NULL,
  alphaCeiling = NULL,
  alphaFloor = NULL,
  alphaInv = NULL,
  alphaMod = NULL,
  alphaModFix = NULL,
  alphaRepl = NULL,
  biLevel = NULL,
  blur = NULL,
  clrChange = NULL,
  clrRepl = NULL,
  duotone = NULL,
  fillOverlay = NULL,
  grayscl = NULL,
  hsl = NULL,
  lum = NULL,
  tint = NULL,
  ...
)
}
\arguments{
\item{cstate}{cstate}

\item{embed}{embed}

\item{link}{link}

\item{noGrp}{noGrp}

\item{noSelect}{noSelect}

\item{noRot}{noRot}

\item{noChangeAspect}{noChangeAspect}

\item{noMove}{noMove}

\item{noResize}{noResize}

\item{noEditPoints}{noEditPoints}

\item{noAdjustHandles}{noAdjustHandles}

\item{noChangeArrowheads}{noChangeArrowheads}

\item{noChangeShapeType}{noChangeShapeType}

\item{extLst}{extLst}

\item{alphaBiLevel}{alphaBiLevel}

\item{alphaCeiling}{alphaCeiling}

\item{alphaFloor}{alphaFloor}

\item{alphaInv}{alphaInv}

\item{alphaMod}{alphaMod}

\item{alphaModFix}{alphaModFix}

\item{alphaRepl}{alphaRepl}

\item{biLevel}{biLevel}

\item{blur}{blur}

\item{clrChange}{clrChange}

\item{clrRepl}{clrRepl}

\item{duotone}{duotone}

\item{fillOverlay}{fillOverlay}

\item{grayscl}{grayscl}

\item{hsl}{hsl}

\item{lum}{lum}

\item{tint}{tint}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
Blip()
}

}
