% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlipFillProperties.R
\name{BlipFillProperties}
\alias{BlipFillProperties}
\title{BlipFillProperties}
\usage{
BlipFillProperties(
  dpi = NULL,
  rotWithShape = NULL,
  blip = NULL,
  tile = NULL,
  srcRect = NULL,
  ...
)
}
\arguments{
\item{dpi}{dpi}

\item{rotWithShape}{rotWithShape}

\item{blip}{blip}

\item{tile}{tile}

\item{srcRect}{srcRect}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
BlipFillProperties()
}

}
