% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlurEffect.R
\name{BlurEffect}
\alias{BlurEffect}
\title{BlurEffect}
\usage{
BlurEffect(rad = NULL, grow = NULL, ...)
}
\arguments{
\item{rad}{rad}

\item{grow}{grow}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
BlurEffect(rad = 0.5)
}

}
