% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BookView.R
\name{BookView}
\alias{BookView}
\title{BookView}
\usage{
BookView(
  visibility = "visible",
  minimized = FALSE,
  showHorizontalScroll = TRUE,
  showVerticalScroll = TRUE,
  showSheetTabs = TRUE,
  xWindow = NULL,
  yWindow = NULL,
  windowWidth = NULL,
  windowHeight = NULL,
  tabRatio = 600L,
  firstSheet = 0L,
  activeTab = 0L,
  autoFilterDateGrouping = TRUE,
  extLst = NULL,
  ...
)
}
\arguments{
\item{visibility}{visibility}

\item{minimized}{minimized}

\item{showHorizontalScroll}{showHorizontalScroll}

\item{showVerticalScroll}{showVerticalScroll}

\item{showSheetTabs}{showSheetTabs}

\item{xWindow}{xWindow}

\item{yWindow}{yWindow}

\item{windowWidth}{windowWidth}

\item{windowHeight}{windowHeight}

\item{tabRatio}{tabRatio}

\item{firstSheet}{firstSheet}

\item{activeTab}{activeTab}

\item{autoFilterDateGrouping}{autoFilterDateGrouping}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.views.html}
}
\examples{
\dontrun{
BookView()
}

}
