% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Border.R
\name{Border}
\alias{Border}
\title{Border}
\usage{
Border(
  left = NULL,
  right = NULL,
  top = NULL,
  bottom = NULL,
  diagonal = NULL,
  diagonal_direction = NULL,
  vertical = NULL,
  horizontal = NULL,
  diagonalUp = FALSE,
  diagonalDown = FALSE,
  outline = TRUE,
  start = NULL,
  end = NULL,
  ...
)
}
\arguments{
\item{left}{left}

\item{right}{right}

\item{top}{top}

\item{bottom}{bottom}

\item{diagonal}{diagonal}

\item{diagonal_direction}{diagonal_direction}

\item{vertical}{vertical}

\item{horizontal}{horizontal}

\item{diagonalUp}{diagonalUp}

\item{diagonalDown}{diagonalDown}

\item{outline}{outline}

\item{start}{start}

\item{end}{end}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Border positioning for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.borders.html}
}
\examples{
\dontrun{
Border()
}

}
