% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Break.R
\name{Break}
\alias{Break}
\title{Break}
\usage{
Break(id = 0L, min = 0L, max = 16383L, man = TRUE, pt = NULL, ...)
}
\arguments{
\item{id}{id}

\item{min}{min}

\item{max}{max}

\item{man}{man}

\item{pt}{pt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.pagebreak.html}
}
\examples{
\dontrun{
Break()
}

}
