% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleChart.R
\name{BubbleChart}
\alias{BubbleChart}
\title{BubbleChart}
\usage{
BubbleChart(
  varyColors = NULL,
  ser = list(),
  dLbls = NULL,
  bubble3D = NULL,
  bubbleScale = NULL,
  showNegBubbles = NULL,
  sizeRepresents = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{varyColors}{varyColors}

\item{ser}{ser}

\item{dLbls}{dLbls}

\item{bubble3D}{bubble3D}

\item{bubbleScale}{bubbleScale}

\item{showNegBubbles}{showNegBubbles}

\item{sizeRepresents}{sizeRepresents}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.bubble_chart.html}
}
\examples{
\dontrun{
BubbleChart()
}

}
