% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_format_code.R
\name{builtin_format_code}
\alias{builtin_format_code}
\title{builtin_format_code}
\usage{
builtin_format_code(index, ...)
}
\arguments{
\item{index}{index}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An character string.
}
\description{
Return one of the standard format codes by index.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.numbers.html}
}
\examples{
\dontrun{
builtin_format_code(index = 2L)
}

}
