% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_format_id.R
\name{builtin_format_id}
\alias{builtin_format_id}
\title{builtin_format_id}
\usage{
builtin_format_id(fmt, ...)
}
\arguments{
\item{fmt}{fmt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Return the id of a standard style.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.numbers.html}
}
\examples{
\dontrun{
builtin_format_id(fmt = NULL)
}

}
