% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcProperties.R
\name{CalcProperties}
\alias{CalcProperties}
\title{CalcProperties}
\usage{
CalcProperties(
  calcId = 124519L,
  calcMode = NULL,
  fullCalcOnLoad = TRUE,
  refMode = NULL,
  iterate = NULL,
  iterateCount = NULL,
  iterateDelta = NULL,
  fullPrecision = NULL,
  calcCompleted = NULL,
  calcOnSave = NULL,
  concurrentCalc = NULL,
  concurrentManualCount = NULL,
  forceFullCalc = NULL,
  ...
)
}
\arguments{
\item{calcId}{calcId}

\item{calcMode}{calcMode}

\item{fullCalcOnLoad}{fullCalcOnLoad}

\item{refMode}{refMode}

\item{iterate}{iterate}

\item{iterateCount}{iterateCount}

\item{iterateDelta}{iterateDelta}

\item{fullPrecision}{fullPrecision}

\item{calcCompleted}{calcCompleted}

\item{calcOnSave}{calcOnSave}

\item{concurrentCalc}{concurrentCalc}

\item{concurrentManualCount}{concurrentManualCount}

\item{forceFullCalc}{forceFullCalc}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.properties.html}
}
\examples{
\dontrun{
CalcProperties()
}

}
