% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Camera.R
\name{Camera}
\alias{Camera}
\title{Camera}
\usage{
Camera(prst = NULL, fov = NULL, zoom = NULL, rot = NULL, ...)
}
\arguments{
\item{prst}{prst}

\item{fov}{fov}

\item{zoom}{zoom}

\item{rot}{rot}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Camera(prst = "perspectiveLeft")
}

}
