% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_numeric.R
\name{cast_numeric}
\alias{cast_numeric}
\title{cast_numeric}
\usage{
cast_numeric(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An numeric value.
}
\description{
Explicitly convert a string to a numeric value

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.inference.html}
}
\examples{
\dontrun{
cast_numeric(value = "1")
}

}
