% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_percentage.R
\name{cast_percentage}
\alias{cast_percentage}
\title{cast_percentage}
\usage{
cast_percentage(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An numeric value.
}
\description{
Explicitly convert a string to numeric value and format as a
percentage

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.inference.html}
}
\examples{
\dontrun{
cast_percentage(value = "50\%")
}

}
