% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_time.R
\name{cast_time}
\alias{cast_time}
\title{cast_time}
\usage{
cast_time(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An Python "datetime.time" object.
}
\description{
Explicitly convert a string to a number and format as datetime or
time

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.inference.html}
}
\examples{
\dontrun{
cast_time(value = "00:00:01")
}

}
