% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{Cell}
\alias{Cell}
\title{Cell}
\usage{
Cell(
  worksheet,
  row = NULL,
  column = NULL,
  value = NULL,
  style_array = NULL,
  ...
)
}
\arguments{
\item{worksheet}{worksheet}

\item{row}{row}

\item{column}{column}

\item{value}{value}

\item{style_array}{style_array}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Describes cell associated properties.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.cell.html}
}
\details{
Properties of interest include style, type, value, and address.
}
\examples{
\dontrun{
wb <- Workbook()
ws <- ws$active
Cell(worksheet = ws, row = 1L, column = 1L)
}

}
