% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellIsRule.R
\name{CellIsRule}
\alias{CellIsRule}
\title{CellIsRule}
\usage{
CellIsRule(
  operator = NULL,
  formula = list(),
  stopIfTrue = NULL,
  font = NULL,
  border = NULL,
  fill = NULL,
  ...
)
}
\arguments{
\item{operator}{operator}

\item{formula}{formula as list}

\item{stopIfTrue}{stopIfTrue}

\item{font}{font}

\item{border}{border}

\item{fill}{fill}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Conditional formatting rule based on cell contents.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
CellIsRule(operator = "equal")
}

}
