% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRange.R
\name{CellRange}
\alias{CellRange}
\title{CellRange}
\usage{
CellRange(
  range_string = NULL,
  min_col = NULL,
  min_row = NULL,
  max_col = NULL,
  max_row = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{range_string}{range_string}

\item{min_col}{min_col}

\item{min_row}{min_row}

\item{max_col}{max_col}

\item{max_row}{max_row}

\item{title}{title}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Represents a range in a sheet: title and coordinates.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.cell_range.html}
}
\details{
This object is used to perform operations on ranges, like: - shift, expand or shrink
- union/intersection with another sheet range, We can check whether a range is: - equal or not equal to another,
- disjoint of another,
- contained in another. We can get: - the size of a range.
- the range bounds (vertices)
- the coordinates,
- the string representation,
}
\examples{
\dontrun{
CellRange(range_string = "A1")
}

}
