% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSmartTag.R
\name{CellSmartTag}
\alias{CellSmartTag}
\title{CellSmartTag}
\usage{
CellSmartTag(
  cellSmartTagPr = list(),
  type = NULL,
  deleted = FALSE,
  xmlBased = FALSE,
  ...
)
}
\arguments{
\item{cellSmartTagPr}{cellSmartTagPr}

\item{type}{type}

\item{deleted}{deleted}

\item{xmlBased}{xmlBased}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.smart_tag.html}
}
\examples{
\dontrun{
CellSmartTag(type = 1L)
}

}
