% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSmartTagPr.R
\name{CellSmartTagPr}
\alias{CellSmartTagPr}
\title{CellSmartTagPr}
\usage{
CellSmartTagPr(key = NULL, val = NULL, ...)
}
\arguments{
\item{key}{key}

\item{val}{val}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.smart_tag.html}
}
\examples{
\dontrun{
CellSmartTagPr(key = "key", val = "val")
}

}
