% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSmartTags.R
\name{CellSmartTags}
\alias{CellSmartTags}
\title{CellSmartTags}
\usage{
CellSmartTags(cellSmartTag = list(), r = NULL, ...)
}
\arguments{
\item{cellSmartTag}{cellSmartTag}

\item{r}{r}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.smart_tag.html}
}
\examples{
\dontrun{
CellSmartTags(cellSmartTag = list(), r = "r")
}

}
