% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStyle.R
\name{CellStyle}
\alias{CellStyle}
\title{CellStyle}
\usage{
CellStyle(
  numFmtId = 0L,
  fontId = 0L,
  fillId = 0L,
  borderId = 0L,
  xfId = NULL,
  quotePrefix = NULL,
  pivotButton = NULL,
  applyNumberFormat = NULL,
  applyFont = NULL,
  applyFill = NULL,
  applyBorder = NULL,
  applyAlignment = NULL,
  applyProtection = NULL,
  alignment = NULL,
  protection = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{numFmtId}{numFmtId}

\item{fontId}{fontId}

\item{fillId}{fillId}

\item{borderId}{borderId}

\item{xfId}{xfId}

\item{quotePrefix}{quotePrefix}

\item{pivotButton}{pivotButton}

\item{applyNumberFormat}{applyNumberFormat}

\item{applyFont}{applyFont}

\item{applyFill}{applyFill}

\item{applyBorder}{applyBorder}

\item{applyAlignment}{applyAlignment}

\item{applyProtection}{applyProtection}

\item{alignment}{alignment}

\item{protection}{protection}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.cell_style.html}
}
\examples{
\dontrun{
CellStyle()
}

}
