% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStyleList.R
\name{CellStyleList}
\alias{CellStyleList}
\title{CellStyleList}
\usage{
CellStyleList(count = NULL, xf = list(), ...)
}
\arguments{
\item{count}{count}

\item{xf}{xf}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.cell_style.html}
}
\examples{
\dontrun{
CellStyleList()
}

}
