% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWatch.R
\name{CellWatch}
\alias{CellWatch}
\title{CellWatch}
\usage{
CellWatch(r = NULL, ...)
}
\arguments{
\item{r}{r}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.cell_watch.html}
}
\examples{
\dontrun{
CellWatch(r = "r")
}

}
