% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWatches.R
\name{CellWatches}
\alias{CellWatches}
\title{CellWatches}
\usage{
CellWatches(cellWatch = list(), ...)
}
\arguments{
\item{cellWatch}{cellWatch}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.cell_watch.html}
}
\examples{
\dontrun{
CellWatches()
}

}
