% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterProperties.R
\name{CharacterProperties}
\alias{CharacterProperties}
\title{CharacterProperties}
\usage{
CharacterProperties(
  kumimoji = NULL,
  lang = NULL,
  altLang = NULL,
  sz = NULL,
  b = NULL,
  i = NULL,
  u = NULL,
  strike = NULL,
  kern = NULL,
  cap = NULL,
  spc = NULL,
  normalizeH = NULL,
  baseline = NULL,
  noProof = NULL,
  dirty = NULL,
  err = NULL,
  smtClean = NULL,
  smtId = NULL,
  bmk = NULL,
  ln = NULL,
  highlight = NULL,
  latin = NULL,
  ea = NULL,
  cs = NULL,
  sym = NULL,
  hlinkClick = NULL,
  hlinkMouseOver = NULL,
  rtl = NULL,
  extLst = NULL,
  noFill = NULL,
  solidFill = NULL,
  gradFill = NULL,
  blipFill = NULL,
  pattFill = NULL,
  grpFill = NULL,
  effectLst = NULL,
  effectDag = NULL,
  uLnTx = NULL,
  uLn = NULL,
  uFillTx = NULL,
  uFill = NULL,
  ...
)
}
\arguments{
\item{kumimoji}{kumimoji}

\item{lang}{lang}

\item{altLang}{altLang}

\item{sz}{sz}

\item{b}{b}

\item{i}{i}

\item{u}{u}

\item{strike}{strike}

\item{kern}{kern}

\item{cap}{cap}

\item{spc}{spc}

\item{normalizeH}{normalizeH}

\item{baseline}{baseline}

\item{noProof}{noProof}

\item{dirty}{dirty}

\item{err}{err}

\item{smtClean}{smtClean}

\item{smtId}{smtId}

\item{bmk}{bmk}

\item{ln}{ln}

\item{highlight}{highlight}

\item{latin}{latin}

\item{ea}{ea}

\item{cs}{cs}

\item{sym}{sym}

\item{hlinkClick}{hlinkClick}

\item{hlinkMouseOver}{hlinkMouseOver}

\item{rtl}{rtl}

\item{extLst}{extLst}

\item{noFill}{noFill}

\item{solidFill}{solidFill}

\item{gradFill}{gradFill}

\item{blipFill}{blipFill}

\item{pattFill}{pattFill}

\item{grpFill}{grpFill}

\item{effectLst}{effectLst}

\item{effectDag}{effectDag}

\item{uLnTx}{uLnTx}

\item{uLn}{uLn}

\item{uFillTx}{uFillTx}

\item{uFill}{uFill}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
CharacterProperties()
}

}
