% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartContainer.R
\name{ChartContainer}
\alias{ChartContainer}
\title{ChartContainer}
\usage{
ChartContainer(
  title = NULL,
  autoTitleDeleted = NULL,
  pivotFmts = list(),
  view3D = NULL,
  floor = NULL,
  sideWall = NULL,
  backWall = NULL,
  plotArea = NULL,
  legend = NULL,
  plotVisOnly = TRUE,
  dispBlanksAs = "gap",
  showDLblsOverMax = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{title}{title}

\item{autoTitleDeleted}{autoTitleDeleted}

\item{pivotFmts}{pivotFmts}

\item{view3D}{view3D}

\item{floor}{floor}

\item{sideWall}{sideWall}

\item{backWall}{backWall}

\item{plotArea}{plotArea}

\item{legend}{legend}

\item{plotVisOnly}{plotVisOnly}

\item{dispBlanksAs}{dispBlanksAs}

\item{showDLblsOverMax}{showDLblsOverMax}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.chartspace.html}
}
\examples{
\dontrun{
ChartContainer()
}

}
