% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartLines.R
\name{ChartLines}
\alias{ChartLines}
\title{ChartLines}
\usage{
ChartLines(spPr = NULL, ...)
}
\arguments{
\item{spPr}{spPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
ChartLines()
}

}
