% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartRelation.R
\name{ChartRelation}
\alias{ChartRelation}
\title{ChartRelation}
\usage{
ChartRelation(id, ...)
}
\arguments{
\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.relation.html}
}
\examples{
\dontrun{
ChartRelation(id = "id")
}

}
