% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chartsheet.R
\name{Chartsheet}
\alias{Chartsheet}
\title{Chartsheet}
\usage{
Chartsheet(
  sheetPr = NULL,
  sheetViews = NULL,
  sheetProtection = NULL,
  customSheetViews = NULL,
  pageMargins = NULL,
  pageSetup = NULL,
  headerFooter = NULL,
  drawing = NULL,
  drawingHF = NULL,
  picture = NULL,
  webPublishItems = NULL,
  extLst = NULL,
  parent = NULL,
  title = "",
  sheet_state = "visible",
  ...
)
}
\arguments{
\item{sheetPr}{sheetPr}

\item{sheetViews}{sheetViews}

\item{sheetProtection}{sheetProtection}

\item{customSheetViews}{customSheetViews}

\item{pageMargins}{pageMargins}

\item{pageSetup}{pageSetup}

\item{headerFooter}{headerFooter}

\item{drawing}{drawing}

\item{drawingHF}{drawingHF}

\item{picture}{picture}

\item{webPublishItems}{webPublishItems}

\item{extLst}{extLst}

\item{parent}{parent}

\item{title}{title}

\item{sheet_state}{sheet_state}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.chartsheet.html}
}
\examples{
\dontrun{
Chartsheet()
}

}
