% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartsheetProperties.R
\name{ChartsheetProperties}
\alias{ChartsheetProperties}
\title{ChartsheetProperties}
\usage{
ChartsheetProperties(published = NULL, codeName = NULL, tabColor = NULL, ...)
}
\arguments{
\item{published}{published}

\item{codeName}{codeName}

\item{tabColor}{tabColor}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.properties.html}
}
\examples{
\dontrun{
ChartsheetProperties()
}

}
