% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartsheetProtection.R
\name{ChartsheetProtection}
\alias{ChartsheetProtection}
\title{ChartsheetProtection}
\usage{
ChartsheetProtection(
  content = NULL,
  objects = NULL,
  hashValue = NULL,
  spinCount = NULL,
  saltValue = NULL,
  algorithmName = NULL,
  password = NULL,
  ...
)
}
\arguments{
\item{content}{content}

\item{objects}{objects}

\item{hashValue}{hashValue}

\item{spinCount}{spinCount}

\item{saltValue}{saltValue}

\item{algorithmName}{algorithmName}

\item{password}{password}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.protection.html}
}
\examples{
\dontrun{
ChartsheetProtection()
}

}
