% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartsheetView.R
\name{ChartsheetView}
\alias{ChartsheetView}
\title{ChartsheetView}
\usage{
ChartsheetView(
  tabSelected = NULL,
  zoomScale = NULL,
  workbookViewId = 0L,
  zoomToFit = TRUE,
  extLst = NULL,
  ...
)
}
\arguments{
\item{tabSelected}{tabSelected}

\item{zoomScale}{zoomScale}

\item{workbookViewId}{workbookViewId}

\item{zoomToFit}{zoomToFit}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.views.html}
}
\examples{
\dontrun{
ChartsheetView()
}

}
