% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartSpace.R
\name{ChartSpace}
\alias{ChartSpace}
\title{ChartSpace}
\usage{
ChartSpace(
  date1904 = NULL,
  lang = NULL,
  roundedCorners = NULL,
  style = NULL,
  clrMapOvr = NULL,
  pivotSource = NULL,
  protection = NULL,
  chart = NULL,
  spPr = NULL,
  txPr = NULL,
  externalData = NULL,
  printSettings = NULL,
  userShapes = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{date1904}{date1904}

\item{lang}{lang}

\item{roundedCorners}{roundedCorners}

\item{style}{style}

\item{clrMapOvr}{clrMapOvr}

\item{pivotSource}{pivotSource}

\item{protection}{protection}

\item{chart}{chart}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{externalData}{externalData}

\item{printSettings}{printSettings}

\item{userShapes}{userShapes}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.chartspace.html}
}
\examples{
\dontrun{
ChartSpace(chart = ChartContainer())
}

}
