% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_to_EMU.R
\name{cm_to_EMU}
\alias{cm_to_EMU}
\title{cm_to_EMU}
\usage{
cm_to_EMU(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
1 cm = 360000 EMUs

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
cm_to_EMU(value = 1L)
}

}
