% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color.R
\name{Color}
\alias{Color}
\title{Color}
\usage{
Color(
  rgb = "00000000",
  indexed = NULL,
  auto = NULL,
  theme = NULL,
  tint = 0,
  index = NULL,
  type = "rgb",
  ...
)
}
\arguments{
\item{rgb}{rgb}

\item{indexed}{indexed}

\item{auto}{auto}

\item{theme}{theme}

\item{tint}{tint}

\item{index}{index}

\item{type}{type}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Named colors for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.colors.html}
}
\examples{
\dontrun{
Color()
}

}
