% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorChangeEffect.R
\name{ColorChangeEffect}
\alias{ColorChangeEffect}
\title{ColorChangeEffect}
\usage{
ColorChangeEffect(useA = NULL, clrFrom = NULL, clrTo = NULL, ...)
}
\arguments{
\item{useA}{useA}

\item{clrFrom}{clrFrom}

\item{clrTo}{clrTo}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
# issue using xlcharts::Color() function as built from
# class 'openpyxl.styles.colors.Color' while
# ColorChangeEffect() arguments require 'openpyxl.drawing.effect.Color'
openpyxl <- reticulate::import("openpyxl")
ColorChangeEffect(
  clrFrom = openpyxl$drawing$effect$Color(),
  clrTo = openpyxl$drawing$effect$Color()
)
}

}
