% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorChoice.R
\name{ColorChoice}
\alias{ColorChoice}
\title{ColorChoice}
\usage{
ColorChoice(
  scrgbClr = NULL,
  srgbClr = NULL,
  hslClr = NULL,
  sysClr = NULL,
  schemeClr = NULL,
  prstClr = NULL,
  ...
)
}
\arguments{
\item{scrgbClr}{scrgbClr}

\item{srgbClr}{srgbClr}

\item{hslClr}{hslClr}

\item{sysClr}{sysClr}

\item{schemeClr}{schemeClr}

\item{prstClr}{prstClr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.colors.html}
}
\examples{
\dontrun{
ColorChoice()
}

}
