% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorFilter.R
\name{ColorFilter}
\alias{ColorFilter}
\title{ColorFilter}
\usage{
ColorFilter(dxfId = NULL, cellColor = NULL, ...)
}
\arguments{
\item{dxfId}{dxfId}

\item{cellColor}{cellColor}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
ColorFilter()
}

}
