% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorList.R
\name{ColorList}
\alias{ColorList}
\title{ColorList}
\usage{
ColorList(indexedColors = list(), mruColors = list(), ...)
}
\arguments{
\item{indexedColors}{indexedColors}

\item{mruColors}{mruColors}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.colors.html}
}
\examples{
\dontrun{
ColorList()
}

}
