% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorMapping.R
\name{ColorMapping}
\alias{ColorMapping}
\title{ColorMapping}
\usage{
ColorMapping(
  bg1 = "lt1",
  tx1 = "dk1",
  bg2 = "lt2",
  tx2 = "dk2",
  accent1 = "accent1",
  accent2 = "accent2",
  accent3 = "accent3",
  accent4 = "accent4",
  accent5 = "accent5",
  accent6 = "accent6",
  hlink = "hlink",
  folHlink = "folHlink",
  extLst = NULL,
  ...
)
}
\arguments{
\item{bg1}{bg1}

\item{tx1}{tx1}

\item{bg2}{bg2}

\item{tx2}{tx2}

\item{accent1}{accent1}

\item{accent2}{accent2}

\item{accent3}{accent3}

\item{accent4}{accent4}

\item{accent5}{accent5}

\item{accent6}{accent6}

\item{hlink}{hlink}

\item{folHlink}{folHlink}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.colors.html}
}
\examples{
\dontrun{
ColorMapping()
}

}
