% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorScale.R
\name{ColorScale}
\alias{ColorScale}
\title{ColorScale}
\usage{
ColorScale(cfvo = NULL, color = NULL, ...)
}
\arguments{
\item{cfvo}{cfvo}

\item{color}{color}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
first <- FormatObject(type = 'min')
last <- FormatObject(type = 'max')
# colors match the format objects:
colors <- list(Color('AA0000'), Color('00AA00'))
cs2 <- ColorScale(cfvo = list(first, last), color = colors)
}

}
