% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorScaleRule.R
\name{ColorScaleRule}
\alias{ColorScaleRule}
\title{ColorScaleRule}
\usage{
ColorScaleRule(
  start_type = NULL,
  start_value = NULL,
  start_color = NULL,
  mid_type = NULL,
  mid_value = NULL,
  mid_color = NULL,
  end_type = NULL,
  end_value = NULL,
  end_color = NULL,
  ...
)
}
\arguments{
\item{start_type}{start_type}

\item{start_value}{start_value}

\item{start_color}{start_color}

\item{mid_type}{mid_type}

\item{mid_value}{mid_value}

\item{mid_color}{mid_color}

\item{end_type}{end_type}

\item{end_value}{end_value}

\item{end_color}{end_color}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Backwards compatibility

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
ColorScaleRule(
start_type='percentile', start_value=10, start_color='FFAA0000',
mid_type='percentile', mid_value=50, mid_color='FF0000AA',
end_type='percentile', end_value=90, end_color='FF00AA00')
}

}
