% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColRange.R
\name{ColRange}
\alias{ColRange}
\title{ColRange}
\usage{
ColRange(range_string = NULL, min_col = NULL, max_col = NULL, ...)
}
\arguments{
\item{range_string}{range_string}

\item{min_col}{min_col}

\item{max_col}{max_col}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Represent a range of at least one column

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.print_settings.html}
}
\examples{
\dontrun{
ColRange(range_string = "A:B")
}

}
