% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_index_from_string.R
\name{column_index_from_string}
\alias{column_index_from_string}
\title{column_index_from_string}
\usage{
column_index_from_string(str_col, ...)
}
\arguments{
\item{str_col}{str_col}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
Convert a column name into a numerical index

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\details{
('A' -> 1)
}
\examples{
\dontrun{
column_index_from_string(str_col = "A")
}

}
