% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnDimension.R
\name{ColumnDimension}
\alias{ColumnDimension}
\title{ColumnDimension}
\usage{
ColumnDimension(
  worksheet,
  index = "A",
  width = 13L,
  bestFit = FALSE,
  hidden = FALSE,
  outlineLevel = 0L,
  outline_level = NULL,
  collapsed = FALSE,
  style = NULL,
  min = NULL,
  max = NULL,
  customWidth = FALSE,
  visible = NULL,
  auto_size = NULL,
  ...
)
}
\arguments{
\item{worksheet}{worksheet}

\item{index}{index}

\item{width}{width}

\item{bestFit}{bestFit}

\item{hidden}{hidden}

\item{outlineLevel}{outlineLevel}

\item{outline_level}{outline_level}

\item{collapsed}{collapsed}

\item{style}{style}

\item{min}{min}

\item{max}{max}

\item{customWidth}{customWidth}

\item{visible}{visible}

\item{auto_size}{auto_size}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Information about the display properties of a column.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.dimensions.html}
}
\examples{
\dontrun{
wb <- Workbook()
ws <- wb$active
ColumnDimension(worksheet = ws)
}

}
