% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comment.R
\name{Comment}
\alias{Comment}
\title{Comment}
\usage{
Comment(text, author, height = 79L, width = 144L, ...)
}
\arguments{
\item{text}{text}

\item{author}{author}

\item{height}{height}

\item{width}{width}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.comments.comments.html}
}
\examples{
\dontrun{
Comment(text = "text", author = "author")
}

}
