% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentRecord.R
\name{CommentRecord}
\alias{CommentRecord}
\title{CommentRecord}
\usage{
CommentRecord(
  ref = "",
  authorId = 0L,
  guid = NULL,
  shapeId = 0L,
  text = NULL,
  commentPr = NULL,
  author = NULL,
  height = 79L,
  width = 144L,
  ...
)
}
\arguments{
\item{ref}{ref}

\item{authorId}{authorId}

\item{guid}{guid}

\item{shapeId}{shapeId}

\item{text}{text}

\item{commentPr}{commentPr}

\item{author}{author}

\item{height}{height}

\item{width}{width}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.comments.comment_sheet.html}
}
\examples{
\dontrun{
CommentRecord()
}

}
