% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormatting.R
\name{ConditionalFormatting}
\alias{ConditionalFormatting}
\title{ConditionalFormatting}
\usage{
ConditionalFormatting(
  sqref = list(),
  pivot = NULL,
  cfRule = list(),
  extLst = NULL,
  ...
)
}
\arguments{
\item{sqref}{sqref}

\item{pivot}{pivot}

\item{cfRule}{cfRule}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.formatting.html}
}
\examples{
\dontrun{
ConditionalFormatting()
}

}
