% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormattingList.R
\name{ConditionalFormattingList}
\alias{ConditionalFormattingList}
\title{ConditionalFormattingList}
\usage{
ConditionalFormattingList(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.formatting.html}
}
\examples{
\dontrun{
ConditionalFormattingList()
}

}
