% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{Connection}
\alias{Connection}
\title{Connection}
\usage{
Connection(id = NULL, idx = NULL, ...)
}
\arguments{
\item{id}{id}

\item{idx}{idx}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.connector.html}
}
\examples{
\dontrun{
Connection(id = 1L, idx = 1L)
}

}
