% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionSite.R
\name{ConnectionSite}
\alias{ConnectionSite}
\title{ConnectionSite}
\usage{
ConnectionSite(ang = NULL, pos = NULL, ...)
}
\arguments{
\item{ang}{ang}

\item{pos}{pos}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
ConnectionSite(ang = 1.5, pos = AdjPoint2D(x = 1L, y = 2L))
}

}
