% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionSiteList.R
\name{ConnectionSiteList}
\alias{ConnectionSiteList}
\title{ConnectionSiteList}
\usage{
ConnectionSiteList(cxn = NULL, ...)
}
\arguments{
\item{cxn}{cxn}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
ConnectionSiteList()
}

}
