% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectorLocking.R
\name{ConnectorLocking}
\alias{ConnectorLocking}
\title{ConnectorLocking}
\usage{
ConnectorLocking(extLst = NULL, ...)
}
\arguments{
\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.connector.html}
}
\examples{
\dontrun{
ConnectorLocking()
}

}
