% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Control.R
\name{Control}
\alias{Control}
\title{Control}
\usage{
Control(controlPr = NULL, shapeId = NULL, name = NULL, ...)
}
\arguments{
\item{controlPr}{controlPr}

\item{shapeId}{shapeId}

\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.

An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.controls.html}
}
\examples{
\dontrun{
Control(
  controlPr = ControlProperty(
    anchor = ObjectAnchor(from = AnchorMarker(), to = AnchorMarker())
    ),
  shapeId = 1L
)
}

}
