% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlProperty.R
\name{ControlProperty}
\alias{ControlProperty}
\title{ControlProperty}
\usage{
ControlProperty(
  anchor = NULL,
  locked = TRUE,
  defaultSize = TRUE,
  print = TRUE,
  disabled = FALSE,
  recalcAlways = FALSE,
  uiObject = FALSE,
  autoFill = TRUE,
  autoLine = TRUE,
  autoPict = TRUE,
  macro = NULL,
  altText = NULL,
  linkedCell = NULL,
  listFillRange = NULL,
  cf = "pict",
  id = NULL,
  ...
)
}
\arguments{
\item{anchor}{anchor}

\item{locked}{locked}

\item{defaultSize}{defaultSize}

\item{print}{_print}

\item{disabled}{disabled}

\item{recalcAlways}{recalcAlways}

\item{uiObject}{uiObject}

\item{autoFill}{autoFill}

\item{autoLine}{autoLine}

\item{autoPict}{autoPict}

\item{macro}{macro}

\item{altText}{altText}

\item{linkedCell}{linkedCell}

\item{listFillRange}{listFillRange}

\item{cf}{cf}

\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.controls.html}
}
\examples{
\dontrun{
ControlProperty(
  anchor = ObjectAnchor(from = AnchorMarker(), to = AnchorMarker())
)
}

}
