% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controls.R
\name{Controls}
\alias{Controls}
\title{Controls}
\usage{
Controls(control = list(), ...)
}
\arguments{
\item{control}{control}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.controls.html}
}
\examples{
\dontrun{
Controls()
}

}
