% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_from_string.R
\name{coordinate_from_string}
\alias{coordinate_from_string}
\title{coordinate_from_string}
\usage{
coordinate_from_string(coord_string, ...)
}
\arguments{
\item{coord_string}{coord_string}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A list.
}
\description{
Convert a coordinate string like 'B12' to a list ('B', 12)

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\examples{
\dontrun{
coordinate_from_string(coord_string = "A1")
}

}
