% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_to_tuple.R
\name{coordinate_to_tuple}
\alias{coordinate_to_tuple}
\title{coordinate_to_tuple}
\usage{
coordinate_to_tuple(coordinate, ...)
}
\arguments{
\item{coordinate}{coordinate}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A list.
}
\description{
Convert an Excel style coordinate to (row, column) list

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\examples{
\dontrun{
coordinate_to_tuple(coordinate = "A1")
}

}
