% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFilter.R
\name{CustomFilter}
\alias{CustomFilter}
\title{CustomFilter}
\usage{
CustomFilter(operator = NULL, val = NULL, ...)
}
\arguments{
\item{operator}{operator}

\item{val}{val}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
CustomFilter(operator = "equal", val = 1L)
}

}
