% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFilters.R
\name{CustomFilters}
\alias{CustomFilters}
\title{CustomFilters}
\usage{
CustomFilters(and = FALSE, customFilter = list(), ...)
}
\arguments{
\item{and}{_and}

\item{customFilter}{customFilter}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
CustomFilters()
}

}
