% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomGeometry2D.R
\name{CustomGeometry2D}
\alias{CustomGeometry2D}
\title{CustomGeometry2D}
\usage{
CustomGeometry2D(
  avLst = NULL,
  gdLst = NULL,
  ahLst = NULL,
  cxnLst = NULL,
  rect = NULL,
  pathLst = NULL,
  ...
)
}
\arguments{
\item{avLst}{avLst}

\item{gdLst}{gdLst}

\item{ahLst}{ahLst}

\item{cxnLst}{cxnLst}

\item{rect}{rect}

\item{pathLst}{pathLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
CustomGeometry2D(pathLst = Path2DList())
}

}
