% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomProperties.R
\name{CustomProperties}
\alias{CustomProperties}
\title{CustomProperties}
\usage{
CustomProperties(customPr = list(), ...)
}
\arguments{
\item{customPr}{customPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.custom.html}
}
\examples{
\dontrun{
CustomProperties()
}

}
