% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomProperty.R
\name{CustomProperty}
\alias{CustomProperty}
\title{CustomProperty}
\usage{
CustomProperty(name = NULL, ...)
}
\arguments{
\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.custom.html}
}
\examples{
\dontrun{
CustomProperty(name = "name")
}

}
