% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomSplit.R
\name{CustomSplit}
\alias{CustomSplit}
\title{CustomSplit}
\usage{
CustomSplit(secondPiePt = list(), ...)
}
\arguments{
\item{secondPiePt}{secondPiePt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pie_chart.html}
}
\examples{
\dontrun{
CustomSplit()
}

}
