% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashStop.R
\name{DashStop}
\alias{DashStop}
\title{DashStop}
\usage{
DashStop(d = 0L, sp = 0L, ...)
}
\arguments{
\item{d}{d}

\item{sp}{sp}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.line.html}
}
\examples{
\dontrun{
DashStop()
}

}
