% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashStopList.R
\name{DashStopList}
\alias{DashStopList}
\title{DashStopList}
\usage{
DashStopList(ds = list(), ...)
}
\arguments{
\item{ds}{ds}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.line.html}
}
\examples{
\dontrun{
DashStopList()
}

}
