% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBar.R
\name{DataBar}
\alias{DataBar}
\title{DataBar}
\usage{
DataBar(
  minLength = NULL,
  maxLength = NULL,
  showValue = NULL,
  cfvo = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{minLength}{minLength}

\item{maxLength}{maxLength}

\item{showValue}{showValue}

\item{cfvo}{cfvo}

\item{color}{color}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
first <- FormatObject(type='min')
second <- FormatObject(type='max')
DataBar(cfvo = list(first, second), color = "638EC6")
}

}
