% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBarRule.R
\name{DataBarRule}
\alias{DataBarRule}
\title{DataBarRule}
\usage{
DataBarRule(
  start_type = NULL,
  start_value = NULL,
  end_type = NULL,
  end_value = NULL,
  color = NULL,
  showValue = NULL,
  minLength = NULL,
  maxLength = NULL,
  ...
)
}
\arguments{
\item{start_type}{start_type}

\item{start_value}{start_value}

\item{end_type}{end_type}

\item{end_value}{end_value}

\item{color}{color}

\item{showValue}{showValue}

\item{minLength}{minLength}

\item{maxLength}{maxLength}

\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
DataBarRule(start_type='percentile', start_value=10, end_type='percentile', end_value='90',
color="FF638EC6", showValue= NULL, minLength=NULL, maxLength=NULL)
}

}
