% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLabel.R
\name{DataLabel}
\alias{DataLabel}
\title{DataLabel}
\usage{
DataLabel(idx = 0L, ...)
}
\arguments{
\item{idx}{idx}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.label.html}
}
\examples{
\dontrun{
DataLabel()
}

}
