% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLabelList.R
\name{DataLabelList}
\alias{DataLabelList}
\title{DataLabelList}
\usage{
DataLabelList(dLbl = list(), delete = NULL, ...)
}
\arguments{
\item{dLbl}{dLbl}

\item{delete}{delete}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.label.html}
}
\examples{
\dontrun{
DataLabelList()
}

}
