% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPoint.R
\name{DataPoint}
\alias{DataPoint}
\title{DataPoint}
\usage{
DataPoint(
  idx = NULL,
  invertIfNegative = NULL,
  marker = NULL,
  bubble3D = NULL,
  explosion = NULL,
  spPr = NULL,
  pictureOptions = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{idx}{idx}

\item{invertIfNegative}{invertIfNegative}

\item{marker}{marker}

\item{bubble3D}{bubble3D}

\item{explosion}{explosion}

\item{spPr}{spPr}

\item{pictureOptions}{pictureOptions}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.marker.html}
}
\examples{
\dontrun{
DataPoint(idx = 1L)
}

}
