% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTable.R
\name{DataTable}
\alias{DataTable}
\title{DataTable}
\usage{
DataTable(
  showHorzBorder = NULL,
  showVertBorder = NULL,
  showOutline = NULL,
  showKeys = NULL,
  spPr = NULL,
  txPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{showHorzBorder}{showHorzBorder}

\item{showVertBorder}{showVertBorder}

\item{showOutline}{showOutline}

\item{showKeys}{showKeys}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.plotarea.html}
}
\examples{
\dontrun{
DataTable()
}

}
