% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTableFormula.R
\name{DataTableFormula}
\alias{DataTableFormula}
\title{DataTableFormula}
\usage{
DataTableFormula(
  ref,
  ca = FALSE,
  dt2D = FALSE,
  dtr = FALSE,
  r1 = NULL,
  r2 = NULL,
  del1 = FALSE,
  del2 = FALSE,
  ...
)
}
\arguments{
\item{ref}{ref}

\item{ca}{ca}

\item{dt2D}{dt2D}

\item{dtr}{dtr}

\item{r1}{r1}

\item{r2}{r2}

\item{del1}{del1}

\item{del2}{del2}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.formula.html}
}
\examples{
\dontrun{
DataTableFormula(ref = "A1")
}

}
