% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataValidation.R
\name{DataValidation}
\alias{DataValidation}
\title{DataValidation}
\usage{
DataValidation(
  type = NULL,
  formula1 = NULL,
  formula2 = NULL,
  showErrorMessage = FALSE,
  showInputMessage = FALSE,
  showDropDown = FALSE,
  allowBlank = FALSE,
  sqref = list(),
  promptTitle = NULL,
  errorStyle = NULL,
  error = NULL,
  prompt = NULL,
  errorTitle = NULL,
  imeMode = NULL,
  operator = NULL,
  allow_blank = FALSE,
  ...
)
}
\arguments{
\item{type}{type}

\item{formula1}{formula1}

\item{formula2}{formula2}

\item{showErrorMessage}{showErrorMessage}

\item{showInputMessage}{showInputMessage}

\item{showDropDown}{showDropDown}

\item{allowBlank}{allowBlank}

\item{sqref}{sqref}

\item{promptTitle}{promptTitle}

\item{errorStyle}{errorStyle}

\item{error}{error}

\item{prompt}{prompt}

\item{errorTitle}{errorTitle}

\item{imeMode}{imeMode}

\item{operator}{operator}

\item{allow_blank}{allow_blank}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.datavalidation.html}
}
\examples{
\dontrun{
DataValidation()
}

}
