% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataValidationList.R
\name{DataValidationList}
\alias{DataValidationList}
\title{DataValidationList}
\usage{
DataValidationList(
  disablePrompts = NULL,
  xWindow = NULL,
  yWindow = NULL,
  count = NULL,
  dataValidation = list(),
  ...
)
}
\arguments{
\item{disablePrompts}{disablePrompts}

\item{xWindow}{xWindow}

\item{yWindow}{yWindow}

\item{count}{count}

\item{dataValidation}{dataValidation}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.datavalidation.html}
}
\examples{
\dontrun{
DataValidationList()
}

}
