% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateAxis.R
\name{DateAxis}
\alias{DateAxis}
\title{DateAxis}
\usage{
DateAxis(
  auto = NULL,
  lblOffset = NULL,
  baseTimeUnit = NULL,
  majorUnit = NULL,
  majorTimeUnit = NULL,
  minorUnit = NULL,
  minorTimeUnit = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{auto}{auto}

\item{lblOffset}{lblOffset}

\item{baseTimeUnit}{baseTimeUnit}

\item{majorUnit}{majorUnit}

\item{majorTimeUnit}{majorTimeUnit}

\item{minorUnit}{minorUnit}

\item{minorTimeUnit}{minorTimeUnit}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
DateAxis()
}

}
