% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateGroupItem.R
\name{DateGroupItem}
\alias{DateGroupItem}
\title{DateGroupItem}
\usage{
DateGroupItem(
  year = NULL,
  month = NULL,
  day = NULL,
  hour = NULL,
  minute = NULL,
  second = NULL,
  dateTimeGrouping = NULL,
  ...
)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{day}{day}

\item{hour}{hour}

\item{minute}{minute}

\item{second}{second}

\item{dateTimeGrouping}{dateTimeGrouping}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
DateGroupItem(year = 2024, dateTimeGrouping = "year")
}

}
