% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcopy.R
\name{deepcopy}
\alias{deepcopy}
\title{Deep copy operation from deepcopy Python module}
\usage{
deepcopy(x, memo = NULL, nil = c())
}
\arguments{
\item{x}{x}

\item{memo}{memo}

\item{nil}{_nil}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://docs.python.org/3/library/copy.html}
}
\examples{
\dontrun{
wb <- Workbook()
deepcopy(wb)
}

}
