% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinedName.R
\name{DefinedName}
\alias{DefinedName}
\title{DefinedName}
\usage{
DefinedName(
  name = NULL,
  comment = NULL,
  customMenu = NULL,
  description = NULL,
  help = NULL,
  statusBar = NULL,
  localSheetId = NULL,
  hidden = NULL,
  vbProcedure = NULL,
  xlm = NULL,
  functionGroupId = NULL,
  shortcutKey = NULL,
  publishToServer = NULL,
  workbookParameter = NULL,
  attr_text = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{comment}{comment}

\item{customMenu}{customMenu}

\item{description}{description}

\item{help}{help}

\item{statusBar}{statusBar}

\item{localSheetId}{localSheetId}

\item{hidden}{hidden}

\item{vbProcedure}{vbProcedure}

\item{xlm}{xlm}

\item{functionGroupId}{functionGroupId}

\item{shortcutKey}{shortcutKey}

\item{publishToServer}{publishToServer}

\item{workbookParameter}{workbookParameter}

\item{attr_text}{attr_text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.defined_name.html}
}
\examples{
\dontrun{
DefinedName(name = "name")
}

}
