% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinedNameList.R
\name{DefinedNameList}
\alias{DefinedNameList}
\title{DefinedNameList}
\usage{
DefinedNameList(definedName = list(), ...)
}
\arguments{
\item{definedName}{definedName}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.defined_name.html}
}
\examples{
\dontrun{
DefinedNameList()
}

}
