% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees_to_angle.R
\name{degrees_to_angle}
\alias{degrees_to_angle}
\title{degrees_to_angle}
\usage{
degrees_to_angle(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
1 degree = 60000 angles

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
degrees_to_angle(value = 1)
}

}
