% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialStyle.R
\name{DifferentialStyle}
\alias{DifferentialStyle}
\title{DifferentialStyle}
\usage{
DifferentialStyle(
  font = NULL,
  numFmt = NULL,
  fill = NULL,
  alignment = NULL,
  border = NULL,
  protection = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{font}{font}

\item{numFmt}{numFmt}

\item{fill}{fill}

\item{alignment}{alignment}

\item{border}{border}

\item{protection}{protection}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.differential.html}
}
\examples{
\dontrun{
DifferentialStyle()
}

}
