% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialStyleList.R
\name{DifferentialStyleList}
\alias{DifferentialStyleList}
\title{DifferentialStyleList}
\usage{
DifferentialStyleList(dxf = list(), count = NULL, ...)
}
\arguments{
\item{dxf}{dxf}

\item{count}{count}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Dedupable container for differential styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.differential.html}
}
\examples{
\dontrun{
DifferentialStyleList()
}

}
