% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension.R
\name{Dimension}
\alias{Dimension}
\title{Dimension}
\usage{
Dimension(
  index,
  hidden,
  outlineLevel,
  collapsed,
  worksheet,
  visible = TRUE,
  style = NULL,
  ...
)
}
\arguments{
\item{index}{index}

\item{hidden}{hidden}

\item{outlineLevel}{outlineLevel}

\item{collapsed}{collapsed}

\item{worksheet}{worksheet}

\item{visible}{visible}

\item{style}{style}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Information about the display properties of a row or column.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.dimensions.html}
}
\examples{
\dontrun{
Dimension(index = 1L, hidden = NULL,
outlineLevel = NULL, collapsed = NULL, worksheet = NULL)
}

}
