% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionHolder.R
\name{DimensionHolder}
\alias{DimensionHolder}
\title{DimensionHolder}
\usage{
DimensionHolder(worksheet, reference = "index", default_factory = NULL, ...)
}
\arguments{
\item{worksheet}{worksheet}

\item{reference}{reference}

\item{default_factory}{default_factory}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Allow columns to be grouped

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.dimensions.html}
}
\examples{
\dontrun{
wb <- Workbook()
ws <- wb$active
DimensionHolder(worksheet = ws)
}

}
