% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayUnitsLabel.R
\name{DisplayUnitsLabel}
\alias{DisplayUnitsLabel}
\title{DisplayUnitsLabel}
\usage{
DisplayUnitsLabel(layout = NULL, tx = NULL, spPr = NULL, txPr = NULL, ...)
}
\arguments{
\item{layout}{layout}

\item{tx}{tx}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
DisplayUnitsLabel()
}

}
