% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentProperties.R
\name{DocumentProperties}
\alias{DocumentProperties}
\title{DocumentProperties}
\usage{
DocumentProperties(
  category = NULL,
  contentStatus = NULL,
  keywords = NULL,
  lastModifiedBy = NULL,
  lastPrinted = NULL,
  revision = NULL,
  version = NULL,
  created = NULL,
  creator = "openpyxl",
  description = NULL,
  identifier = NULL,
  language = NULL,
  modified = NULL,
  subject = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{category}{category}

\item{contentStatus}{contentStatus}

\item{keywords}{keywords}

\item{lastModifiedBy}{lastModifiedBy}

\item{lastPrinted}{lastPrinted}

\item{revision}{revision}

\item{version}{version}

\item{created}{created}

\item{creator}{creator}

\item{description}{description}

\item{identifier}{identifier}

\item{language}{language}

\item{modified}{modified}

\item{subject}{subject}

\item{title}{title}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
High-level properties of the document.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.packaging.core.html}
}
\details{
Defined in ECMA-376 Par2 Annex D
}
\examples{
\dontrun{
DocumentProperties()
}

}
