% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoughnutChart.R
\name{DoughnutChart}
\alias{DoughnutChart}
\title{DoughnutChart}
\usage{
DoughnutChart(firstSliceAng = 0L, holeSize = 10L, extLst = NULL, ...)
}
\arguments{
\item{firstSliceAng}{firstSliceAng}

\item{holeSize}{holeSize}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pie_chart.html}
}
\examples{
\dontrun{
DoughnutChart()
}

}
