% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drawing.R
\name{Drawing}
\alias{Drawing}
\title{Drawing}
\usage{
Drawing(id = NULL, ...)
}
\arguments{
\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.drawing.html}
}
\examples{
\dontrun{
Drawing()
}

}
