% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingHF.R
\name{DrawingHF}
\alias{DrawingHF}
\title{DrawingHF}
\usage{
DrawingHF(
  id = NULL,
  lho = NULL,
  lhe = NULL,
  lhf = NULL,
  cho = NULL,
  che = NULL,
  chf = NULL,
  rho = NULL,
  rhe = NULL,
  rhf = NULL,
  lfo = NULL,
  lfe = NULL,
  lff = NULL,
  cfo = NULL,
  cfe = NULL,
  cff = NULL,
  rfo = NULL,
  rfe = NULL,
  rff = NULL,
  ...
)
}
\arguments{
\item{id}{id}

\item{lho}{lho}

\item{lhe}{lhe}

\item{lhf}{lhf}

\item{cho}{cho}

\item{che}{che}

\item{chf}{chf}

\item{rho}{rho}

\item{rhe}{rhe}

\item{rhf}{rhf}

\item{lfo}{lfo}

\item{lfe}{lfe}

\item{lff}{lff}

\item{cfo}{cfo}

\item{cfe}{cfe}

\item{cff}{cff}

\item{rfo}{rfo}

\item{rfe}{rfe}

\item{rff}{rff}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.relation.html}
}
\examples{
\dontrun{
DrawingHF()
}

}
