% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyWorksheet.R
\name{DummyWorksheet}
\alias{DummyWorksheet}
\title{DummyWorksheet}
\usage{
DummyWorksheet(title, ...)
}
\arguments{
\item{title}{title}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.reference.html}
}
\examples{
\dontrun{
DummyWorksheet(title = "title")
}

}
