% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuotoneEffect.R
\name{DuotoneEffect}
\alias{DuotoneEffect}
\title{DuotoneEffect}
\usage{
DuotoneEffect(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
DuotoneEffect()
}

}
