% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicFilter.R
\name{DynamicFilter}
\alias{DynamicFilter}
\title{DynamicFilter}
\usage{
DynamicFilter(
  type = NULL,
  val = NULL,
  valIso = NULL,
  maxVal = NULL,
  maxValIso = NULL,
  ...
)
}
\arguments{
\item{type}{type}

\item{val}{val}

\item{valIso}{valIso}

\item{maxVal}{maxVal}

\item{maxValIso}{maxValIso}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
DynamicFilter(type = "thisWeek")
}

}
