% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectContainer.R
\name{EffectContainer}
\alias{EffectContainer}
\title{EffectContainer}
\usage{
EffectContainer(type = NULL, name = NULL, ...)
}
\arguments{
\item{type}{type}

\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
EffectContainer(type = "sib")
}

}
