% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectList.R
\name{EffectList}
\alias{EffectList}
\title{EffectList}
\usage{
EffectList(
  blur = NULL,
  fillOverlay = NULL,
  glow = NULL,
  innerShdw = NULL,
  outerShdw = NULL,
  prstShdw = NULL,
  reflection = NULL,
  softEdge = NULL,
  ...
)
}
\arguments{
\item{blur}{blur}

\item{fillOverlay}{fillOverlay}

\item{glow}{glow}

\item{innerShdw}{innerShdw}

\item{outerShdw}{outerShdw}

\item{prstShdw}{prstShdw}

\item{reflection}{reflection}

\item{softEdge}{softEdge}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
EffectList()
}

}
