% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbeddedWAVAudioFile.R
\name{EmbeddedWAVAudioFile}
\alias{EmbeddedWAVAudioFile}
\title{EmbeddedWAVAudioFile}
\usage{
EmbeddedWAVAudioFile(name = NULL, ...)
}
\arguments{
\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
EmbeddedWAVAudioFile()
}

}
