% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyCell.R
\name{EmptyCell}
\alias{EmptyCell}
\title{EmptyCell}
\usage{
EmptyCell(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.read_only.html}
}
\examples{
\dontrun{
EmptyCell()
}

}
