% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMU_to_cm.R
\name{EMU_to_cm}
\alias{EMU_to_cm}
\title{EMU_to_cm}
\usage{
EMU_to_cm(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
EMU_to_cm(value = 100L)
}

}
