% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorBars.R
\name{ErrorBars}
\alias{ErrorBars}
\title{ErrorBars}
\usage{
ErrorBars(
  errDir = NULL,
  errBarType = "both",
  errValType = "fixedVal",
  noEndCap = NULL,
  plus = NULL,
  minus = NULL,
  val = NULL,
  spPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{errDir}{errDir}

\item{errBarType}{errBarType}

\item{errValType}{errValType}

\item{noEndCap}{noEndCap}

\item{plus}{plus}

\item{minus}{minus}

\item{val}{val}

\item{spPr}{spPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.error_bar.html}
}
\examples{
\dontrun{
ErrorBars()
}

}
