% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelReader.R
\name{ExcelReader}
\alias{ExcelReader}
\title{ExcelReader}
\usage{
ExcelReader(
  fn,
  read_only = FALSE,
  keep_vba = FALSE,
  data_only = FALSE,
  keep_links = TRUE,
  rich_text = FALSE,
  ...
)
}
\arguments{
\item{fn}{fn}

\item{read_only}{read_only}

\item{keep_vba}{keep_vba}

\item{data_only}{data_only}

\item{keep_links}{keep_links}

\item{rich_text}{rich_text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Read an Excel package and dispatch the contents to the relevant modules

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.reader.excel.html}
}
\examples{
\dontrun{
tmp <- write_xlsx(list(mysheet = iris))
ExcelReader(fn = tmp)
}

}
