% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelWriter.R
\name{ExcelWriter}
\alias{ExcelWriter}
\title{ExcelWriter}
\usage{
ExcelWriter(workbook, archive, ...)
}
\arguments{
\item{workbook}{workbook}

\item{archive}{archive}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Write a workbook object to an Excel file.

\url{https://openpyxl.readthedocs.io/en/latest/api/openpyxl.writer.excel.html}
}
\examples{
\dontrun{
wb <- Workbook()
ExcelWriter(workbook = wb, archive = wb)
}

}
