% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extension.R
\name{Extension}
\alias{Extension}
\title{Extension}
\usage{
Extension(uri = NULL, ...)
}
\arguments{
\item{uri}{uri}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.errors.html}
}
\examples{
\dontrun{
Extension()
}

}
