% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtensionList.R
\name{ExtensionList}
\alias{ExtensionList}
\title{ExtensionList}
\usage{
ExtensionList(ext = list(), ...)
}
\arguments{
\item{ext}{ext}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.descriptors.excel.html}

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.errors.html}
}
\examples{
\dontrun{
ExtensionList()
}

}
