% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalBook.R
\name{ExternalBook}
\alias{ExternalBook}
\title{ExternalBook}
\usage{
ExternalBook(
  sheetNames = NULL,
  definedNames = list(),
  sheetDataSet = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{sheetNames}{sheetNames}

\item{definedNames}{definedNames}

\item{sheetDataSet}{sheetDataSet}

\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalBook()
}

}
