% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalCell.R
\name{ExternalCell}
\alias{ExternalCell}
\title{ExternalCell}
\usage{
ExternalCell(r = NULL, t = NULL, vm = NULL, v = NULL, ...)
}
\arguments{
\item{r}{r}

\item{t}{t}

\item{vm}{vm}

\item{v}{v}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalCell(r = "r")
}

}
