% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalData.R
\name{ExternalData}
\alias{ExternalData}
\title{ExternalData}
\usage{
ExternalData(autoUpdate = NULL, id = NULL, ...)
}
\arguments{
\item{autoUpdate}{autoUpdate}

\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.chartspace.html}
}
\examples{
\dontrun{
ExternalData(id = "id")
}

}
