% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalDefinedName.R
\name{ExternalDefinedName}
\alias{ExternalDefinedName}
\title{ExternalDefinedName}
\usage{
ExternalDefinedName(name = NULL, refersTo = NULL, sheetId = NULL, ...)
}
\arguments{
\item{name}{name}

\item{refersTo}{refersTo}

\item{sheetId}{sheetId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalDefinedName(name = "name")
}

}
