% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalLink.R
\name{ExternalLink}
\alias{ExternalLink}
\title{ExternalLink}
\usage{
ExternalLink(
  externalBook = NULL,
  ddeLink = NULL,
  oleLink = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{externalBook}{externalBook}

\item{ddeLink}{ddeLink}

\item{oleLink}{oleLink}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalLink()
}

}
