% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalReference.R
\name{ExternalReference}
\alias{ExternalReference}
\title{ExternalReference}
\usage{
ExternalReference(id, ...)
}
\arguments{
\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_reference.html}
}
\examples{
\dontrun{
ExternalReference(id = "id")
}

}
