% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalSheetDataSet.R
\name{ExternalSheetDataSet}
\alias{ExternalSheetDataSet}
\title{ExternalSheetDataSet}
\usage{
ExternalSheetDataSet(sheetData = NULL, ...)
}
\arguments{
\item{sheetData}{sheetData}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalSheetDataSet(sheetData = list())
}

}
