% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalSheetNames.R
\name{ExternalSheetNames}
\alias{ExternalSheetNames}
\title{ExternalSheetNames}
\usage{
ExternalSheetNames(sheetName = list(), ...)
}
\arguments{
\item{sheetName}{sheetName}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.external_link.external.html}
}
\examples{
\dontrun{
ExternalSheetNames()
}

}
