% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSharing.R
\name{FileSharing}
\alias{FileSharing}
\title{FileSharing}
\usage{
FileSharing(
  readOnlyRecommended = NULL,
  userName = NULL,
  reservationPassword = NULL,
  algorithmName = NULL,
  hashValue = NULL,
  saltValue = NULL,
  spinCount = NULL,
  ...
)
}
\arguments{
\item{readOnlyRecommended}{readOnlyRecommended}

\item{userName}{userName}

\item{reservationPassword}{reservationPassword}

\item{algorithmName}{algorithmName}

\item{hashValue}{hashValue}

\item{saltValue}{saltValue}

\item{spinCount}{spinCount}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.protection.html}
}
\examples{
\dontrun{
FileSharing()
}

}
