% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileVersion.R
\name{FileVersion}
\alias{FileVersion}
\title{FileVersion}
\usage{
FileVersion(
  appName = NULL,
  lastEdited = NULL,
  lowestEdited = NULL,
  rupBuild = NULL,
  codeName = NULL,
  ...
)
}
\arguments{
\item{appName}{appName}

\item{lastEdited}{lastEdited}

\item{lowestEdited}{lowestEdited}

\item{rupBuild}{rupBuild}

\item{codeName}{codeName}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.properties.html}
}
\examples{
\dontrun{
FileVersion()
}

}
