% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterColumn.R
\name{FilterColumn}
\alias{FilterColumn}
\title{FilterColumn}
\usage{
FilterColumn(
  colId = NULL,
  hiddenButton = FALSE,
  showButton = TRUE,
  filters = NULL,
  top10 = NULL,
  customFilters = NULL,
  dynamicFilter = NULL,
  colorFilter = NULL,
  iconFilter = NULL,
  extLst = NULL,
  blank = NULL,
  vals = NULL,
  ...
)
}
\arguments{
\item{colId}{colId}

\item{hiddenButton}{hiddenButton}

\item{showButton}{showButton}

\item{filters}{filters}

\item{top10}{top10}

\item{customFilters}{customFilters}

\item{dynamicFilter}{dynamicFilter}

\item{colorFilter}{colorFilter}

\item{iconFilter}{iconFilter}

\item{extLst}{extLst}

\item{blank}{blank}

\item{vals}{vals}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
FilterColumn(colId = 1L)
}

}
