% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Filters}
\alias{Filters}
\title{Filters}
\usage{
Filters(
  blank = NULL,
  calendarType = NULL,
  filter = list(),
  dateGroupItem = list(),
  ...
)
}
\arguments{
\item{blank}{blank}

\item{calendarType}{calendarType}

\item{filter}{filter}

\item{dateGroupItem}{dateGroupItem}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
Filters()
}

}
