% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Font.R
\name{Font}
\alias{Font}
\title{Font}
\usage{
Font(
  name = NULL,
  sz = NULL,
  b = NULL,
  i = NULL,
  charset = NULL,
  u = NULL,
  strike = NULL,
  color = NULL,
  scheme = NULL,
  family = NULL,
  size = NULL,
  bold = NULL,
  italic = NULL,
  strikethrough = NULL,
  underline = NULL,
  vertAlign = NULL,
  outline = NULL,
  shadow = NULL,
  condense = NULL,
  extend = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{sz}{sz}

\item{b}{b}

\item{i}{i}

\item{charset}{charset}

\item{u}{u}

\item{strike}{strike}

\item{color}{color}

\item{scheme}{scheme}

\item{family}{family}

\item{size}{size}

\item{bold}{bold}

\item{italic}{italic}

\item{strikethrough}{strikethrough}

\item{underline}{underline}

\item{vertAlign}{vertAlign}

\item{outline}{outline}

\item{shadow}{shadow}

\item{condense}{condense}

\item{extend}{extend}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Font options used in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.fonts.html}
}
\examples{
\dontrun{
Font()
}

}
