% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FontReference.R
\name{FontReference}
\alias{FontReference}
\title{FontReference}
\usage{
FontReference(idx = NULL, ...)
}
\arguments{
\item{idx}{idx}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
FontReference()
}

}
