% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatObject.R
\name{FormatObject}
\alias{FormatObject}
\title{FormatObject}
\usage{
FormatObject(type, val = NULL, gte = NULL, extLst = NULL, ...)
}
\arguments{
\item{type}{type}

\item{val}{val}

\item{gte}{gte}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
FormatObject(type = "num")
}

}
