% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaRule.R
\name{FormulaRule}
\alias{FormulaRule}
\title{FormulaRule}
\usage{
FormulaRule(
  formula = NULL,
  stopIfTrue = NULL,
  font = NULL,
  border = NULL,
  fill = NULL,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{stopIfTrue}{stopIfTrue}

\item{font}{font}

\item{border}{border}

\item{fill}{fill}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Conditional formatting with custom differential style

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
wb <- Workbook()
ws <- wb$active
ws$conditional_formatting$add('E1:E10',
  FormulaRule(
    formula = list('ISBLANK(A1)'),
    stopIfTrue=TRUE
  )
)
}

}
