% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGroup.R
\name{FunctionGroup}
\alias{FunctionGroup}
\title{FunctionGroup}
\usage{
FunctionGroup(name = NULL, ...)
}
\arguments{
\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.function_group.html}
}
\examples{
\dontrun{
FunctionGroup(name = "name")
}

}
