% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGroupList.R
\name{FunctionGroupList}
\alias{FunctionGroupList}
\title{FunctionGroupList}
\usage{
FunctionGroupList(builtInGroupCount = 16L, functionGroup = list(), ...)
}
\arguments{
\item{builtInGroupCount}{builtInGroupCount}

\item{functionGroup}{functionGroup}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.function_group.html}
}
\examples{
\dontrun{
FunctionGroupList()
}

}
