% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomGuide.R
\name{GeomGuide}
\alias{GeomGuide}
\title{GeomGuide}
\usage{
GeomGuide(name = NULL, fmla = NULL, ...)
}
\arguments{
\item{name}{name}

\item{fmla}{fmla}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
GeomGuide(name = "name", fmla = "fmla")
}

}
