% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomGuideList.R
\name{GeomGuideList}
\alias{GeomGuideList}
\title{GeomGuideList}
\usage{
GeomGuideList(gd = NULL, ...)
}
\arguments{
\item{gd}{gd}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
GeomGuideList()
}

}
