% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomRect.R
\name{GeomRect}
\alias{GeomRect}
\title{GeomRect}
\usage{
GeomRect(l = NULL, t = NULL, r = NULL, b = NULL, ...)
}
\arguments{
\item{l}{l}

\item{t}{t}

\item{r}{r}

\item{b}{b}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
GeomRect(l = 1L, t = 1L, 1L, 1L)
}

}
