% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_letter.R
\name{get_column_letter}
\alias{get_column_letter}
\title{get_column_letter}
\usage{
get_column_letter(idx, ...)
}
\arguments{
\item{idx}{idx}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Convert a column index into a column letter

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\details{
(3 -> 'C')
}
\examples{
\dontrun{
get_column_letter(idx = 3L)
}

}
