% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type.R
\name{get_type}
\alias{get_type}
\title{get_type}
\usage{
get_type(t, value, ...)
}
\arguments{
\item{t}{t}

\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.cell.html}
}
\examples{
\dontrun{
get_type(t = "t", value = "value")
}

}
