% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlowEffect.R
\name{GlowEffect}
\alias{GlowEffect}
\title{GlowEffect}
\usage{
GlowEffect(rad = NULL, ...)
}
\arguments{
\item{rad}{rad}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
GlowEffect(rad = 1.5)
}

}
