% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientFillProperties.R
\name{GradientFillProperties}
\alias{GradientFillProperties}
\title{GradientFillProperties}
\usage{
GradientFillProperties(
  flip = NULL,
  rotWithShape = NULL,
  gsLst = list(),
  lin = NULL,
  path = NULL,
  tileRect = NULL,
  ...
)
}
\arguments{
\item{flip}{flip}

\item{rotWithShape}{rotWithShape}

\item{gsLst}{gsLst}

\item{lin}{lin}

\item{path}{path}

\item{tileRect}{tileRect}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
GradientFillProperties()
}

}
