% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientStop.R
\name{GradientStop}
\alias{GradientStop}
\title{GradientStop}
\usage{
GradientStop(
  pos = NULL,
  scrgbClr = NULL,
  srgbClr = NULL,
  hslClr = NULL,
  sysClr = NULL,
  schemeClr = NULL,
  prstClr = NULL,
  ...
)
}
\arguments{
\item{pos}{pos}

\item{scrgbClr}{scrgbClr}

\item{srgbClr}{srgbClr}

\item{hslClr}{hslClr}

\item{sysClr}{sysClr}

\item{schemeClr}{schemeClr}

\item{prstClr}{prstClr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
GradientStop()
}

}
