% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalProperties.R
\name{GraphicalProperties}
\alias{GraphicalProperties}
\title{GraphicalProperties}
\usage{
GraphicalProperties(
  bwMode = NULL,
  xfrm = NULL,
  noFill = NULL,
  solidFill = NULL,
  gradFill = NULL,
  pattFill = NULL,
  ln = NULL,
  scene3d = NULL,
  custGeom = NULL,
  prstGeom = NULL,
  sp3d = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{bwMode}{bwMode}

\item{xfrm}{xfrm}

\item{noFill}{noFill}

\item{solidFill}{solidFill}

\item{gradFill}{gradFill}

\item{pattFill}{pattFill}

\item{ln}{ln}

\item{scene3d}{scene3d}

\item{custGeom}{custGeom}

\item{prstGeom}{prstGeom}

\item{sp3d}{sp3d}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.shapes.html}
}
\details{
This element specifies the formatting for the parent chart element. The
custGeom, prstGeom, scene3d, and xfrm elements are not supported. The
bwMode attribute is not supported. This doesn't leave much. And the element is used in different places.
}
\examples{
\dontrun{
GraphicalProperties()
}

}
