% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicData.R
\name{GraphicData}
\alias{GraphicData}
\title{GraphicData}
\usage{
GraphicData(
  uri = "http://schemas.openxmlformats.org/drawingml/2006/chart",
  chart = NULL,
  ...
)
}
\arguments{
\item{uri}{uri}

\item{chart}{chart}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.graphic.html}
}
\examples{
\dontrun{
GraphicData()
}

}
