% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFrameLocking.R
\name{GraphicFrameLocking}
\alias{GraphicFrameLocking}
\title{GraphicFrameLocking}
\usage{
GraphicFrameLocking(
  noGrp = NULL,
  noDrilldown = NULL,
  noSelect = NULL,
  noChangeAspect = NULL,
  noMove = NULL,
  noResize = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{noGrp}{noGrp}

\item{noDrilldown}{noDrilldown}

\item{noSelect}{noSelect}

\item{noChangeAspect}{noChangeAspect}

\item{noMove}{noMove}

\item{noResize}{noResize}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.graphic.html}
}
\examples{
\dontrun{
GraphicFrameLocking()
}

}
