% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLocking.R
\name{GroupLocking}
\alias{GroupLocking}
\title{GroupLocking}
\usage{
GroupLocking(
  noGrp = NULL,
  noUngrp = NULL,
  noSelect = NULL,
  noRot = NULL,
  noChangeAspect = NULL,
  noChangeArrowheads = NULL,
  noMove = NULL,
  noResize = NULL,
  noEditPoints = NULL,
  noAdjustHandles = NULL,
  noChangeShapeType = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{noGrp}{noGrp}

\item{noUngrp}{noUngrp}

\item{noSelect}{noSelect}

\item{noRot}{noRot}

\item{noChangeAspect}{noChangeAspect}

\item{noChangeArrowheads}{noChangeArrowheads}

\item{noMove}{noMove}

\item{noResize}{noResize}

\item{noEditPoints}{noEditPoints}

\item{noAdjustHandles}{noAdjustHandles}

\item{noChangeShapeType}{noChangeShapeType}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.properties.html}
}
\examples{
\dontrun{
GroupLocking()
}

}
