% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupShapeProperties.R
\name{GroupShapeProperties}
\alias{GroupShapeProperties}
\title{GroupShapeProperties}
\usage{
GroupShapeProperties(
  bwMode = NULL,
  xfrm = NULL,
  scene3d = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{bwMode}{bwMode}

\item{xfrm}{xfrm}

\item{scene3d}{scene3d}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.properties.html}
}
\examples{
\dontrun{
GroupShapeProperties()
}

}
