% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupTransform2D.R
\name{GroupTransform2D}
\alias{GroupTransform2D}
\title{GroupTransform2D}
\usage{
GroupTransform2D(
  rot = 0L,
  flipH = NULL,
  flipV = NULL,
  off = NULL,
  ext = NULL,
  chOff = NULL,
  chExt = NULL,
  ...
)
}
\arguments{
\item{rot}{rot}

\item{flipH}{flipH}

\item{flipV}{flipV}

\item{off}{off}

\item{ext}{ext}

\item{chOff}{chOff}

\item{chExt}{chExt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
GroupTransform2D()
}

}
