% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_password.R
\name{hash_password}
\alias{hash_password}
\title{hash_password}
\usage{
hash_password(plaintext_password = "", ...)
}
\arguments{
\item{plaintext_password}{plaintext_password}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Create a password hash from a given string for protecting a worksheet

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.protection.html}
}
\details{
only. This will not work for encrypting a workbook. This method is based on the algorithm provided by
Daniel Rentz of OpenOffice and the PEAR package
Spreadsheet_Excel_Writer by Xavier Noguer <xnoguer@rezebra.com>.
See also http://blogs.msdn.com/b/ericwhite/archive/2008/02/23/the-legacy-hashing-algorithm-in-open-xml.aspx
}
\examples{
\dontrun{
hash_password(plaintext_password = "password")
}

}
