% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeaderFooter.R
\name{HeaderFooter}
\alias{HeaderFooter}
\title{HeaderFooter}
\usage{
HeaderFooter(
  differentOddEven = NULL,
  differentFirst = NULL,
  scaleWithDoc = NULL,
  alignWithMargins = NULL,
  oddHeader = NULL,
  oddFooter = NULL,
  evenHeader = NULL,
  evenFooter = NULL,
  firstHeader = NULL,
  firstFooter = NULL,
  ...
)
}
\arguments{
\item{differentOddEven}{differentOddEven}

\item{differentFirst}{differentFirst}

\item{scaleWithDoc}{scaleWithDoc}

\item{alignWithMargins}{alignWithMargins}

\item{oddHeader}{oddHeader}

\item{oddFooter}{oddFooter}

\item{evenHeader}{evenHeader}

\item{evenFooter}{evenFooter}

\item{firstHeader}{firstHeader}

\item{firstFooter}{firstFooter}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.header_footer.html}
}
\examples{
\dontrun{
HeaderFooter()
}

}
