% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeaderFooterItem.R
\name{HeaderFooterItem}
\alias{HeaderFooterItem}
\title{HeaderFooterItem}
\usage{
HeaderFooterItem(left = NULL, right = NULL, center = NULL, ...)
}
\arguments{
\item{left}{left}

\item{right}{right}

\item{center}{center}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Header or footer item

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.header_footer.html}
}
\examples{
\dontrun{
HeaderFooterItem()
}

}
