% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSLColor.R
\name{HSLColor}
\alias{HSLColor}
\title{HSLColor}
\usage{
HSLColor(hue = NULL, sat = NULL, lum = NULL, ...)
}
\arguments{
\item{hue}{hue}

\item{sat}{sat}

\item{lum}{lum}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.colors.html}
}
\examples{
\dontrun{
HSLColor(hue = 1L, sat = 1L, lum = 1L)
}

}
