% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSLEffect.R
\name{HSLEffect}
\alias{HSLEffect}
\title{HSLEffect}
\usage{
HSLEffect(hue = NULL, sat = NULL, lum = NULL, ...)
}
\arguments{
\item{hue}{hue}

\item{sat}{sat}

\item{lum}{lum}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
HSLEffect(hue = 1L, sat = 1L, lum = 1L)
}

}
