% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperlink.R
\name{Hyperlink}
\alias{Hyperlink}
\title{Hyperlink}
\usage{
Hyperlink(
  ref = NULL,
  location = NULL,
  tooltip = NULL,
  display = NULL,
  id = NULL,
  target = NULL,
  ...
)
}
\arguments{
\item{ref}{ref}

\item{location}{location}

\item{tooltip}{tooltip}

\item{display}{display}

\item{id}{id}

\item{target}{target}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.hyperlink.html}
}
\examples{
\dontrun{
Hyperlink(ref = "ref")
}

}
