% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperlinkList.R
\name{HyperlinkList}
\alias{HyperlinkList}
\title{HyperlinkList}
\usage{
HyperlinkList(hyperlink = list(), ...)
}
\arguments{
\item{hyperlink}{hyperlink}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.hyperlink.html}
}
\examples{
\dontrun{
HyperlinkList()
}

}
