% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IconFilter.R
\name{IconFilter}
\alias{IconFilter}
\title{IconFilter}
\usage{
IconFilter(iconSet = NULL, iconId = NULL, ...)
}
\arguments{
\item{iconSet}{iconSet}

\item{iconId}{iconId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
IconFilter(iconSet = "3TrafficLights1")
}

}
