% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IconSet.R
\name{IconSet}
\alias{IconSet}
\title{IconSet}
\usage{
IconSet(
  iconSet = NULL,
  showValue = NULL,
  percent = NULL,
  reverse = NULL,
  cfvo = NULL,
  ...
)
}
\arguments{
\item{iconSet}{iconSet}

\item{showValue}{showValue}

\item{percent}{percent}

\item{reverse}{reverse}

\item{cfvo}{cfvo}

\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
first <- FormatObject(type='percent', val=0)
second <- FormatObject(type='percent', val=33)
third <- FormatObject(type='percent', val=67)
iconset <- IconSet(
  iconSet = '3TrafficLights1',
  cfvo = list(first, second, third)
)
}

}
