% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IconSetRule.R
\name{IconSetRule}
\alias{IconSetRule}
\title{IconSetRule}
\usage{
IconSetRule(
  icon_style = NULL,
  type = NULL,
  values = NULL,
  showValue = NULL,
  percent = NULL,
  reverse = NULL,
  ...
)
}
\arguments{
\item{icon_style}{icon_style}

\item{type}{type}

\item{values}{values}

\item{showValue}{showValue}

\item{percent}{percent}

\item{reverse}{reverse}

\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
Convenience function for creating icon set rules

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
IconSetRule('5Arrows', 'percent', list(10, 20, 30, 40, 50),
showValue=NULL, percent=NULL, reverse=NULL)
}
}
