% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgnoredError.R
\name{IgnoredError}
\alias{IgnoredError}
\title{IgnoredError}
\usage{
IgnoredError(
  sqref = NULL,
  evalError = FALSE,
  twoDigitTextYear = FALSE,
  numberStoredAsText = FALSE,
  formula = FALSE,
  formulaRange = FALSE,
  unlockedFormula = FALSE,
  emptyCellReference = FALSE,
  listDataValidation = FALSE,
  calculatedColumn = FALSE,
  ...
)
}
\arguments{
\item{sqref}{sqref}

\item{evalError}{evalError}

\item{twoDigitTextYear}{twoDigitTextYear}

\item{numberStoredAsText}{numberStoredAsText}

\item{formula}{formula}

\item{formulaRange}{formulaRange}

\item{unlockedFormula}{unlockedFormula}

\item{emptyCellReference}{emptyCellReference}

\item{listDataValidation}{listDataValidation}

\item{calculatedColumn}{calculatedColumn}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.errors.html}
}
\examples{
\dontrun{
IgnoredError()
}
}
