% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgnoredErrors.R
\name{IgnoredErrors}
\alias{IgnoredErrors}
\title{IgnoredErrors}
\usage{
IgnoredErrors(ignoredError = list(), extLst = NULL, ...)
}
\arguments{
\item{ignoredError}{ignoredError}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.errors.html}
}
\examples{
\dontrun{
IgnoredErrors()
}
}
