% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{Image}
\alias{Image}
\title{Image}
\usage{
Image(img, ...)
}
\arguments{
\item{img}{img}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Image in a spreadsheet. You must have the Pillow Python library
installed, which can be installed using `install_pillow()`.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.image.html}
}
\examples{
\dontrun{
install_pillow()
image_path <- system.file('extdata/logo.png',package='xlcharts')
Image(img = image_path)
}
}
