% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch_to_EMU.R
\name{inch_to_EMU}
\alias{inch_to_EMU}
\title{inch_to_EMU}
\usage{
inch_to_EMU(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An numeric value.
}
\description{
1 inch = 914400 EMUs

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
inch_to_EMU(value = 1)
}
}
