% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InlineFont.R
\name{InlineFont}
\alias{InlineFont}
\title{InlineFont}
\usage{
InlineFont(
  rFont = NULL,
  charset = NULL,
  family = NULL,
  b = NULL,
  i = NULL,
  strike = NULL,
  outline = NULL,
  shadow = NULL,
  condense = NULL,
  extend = NULL,
  color = NULL,
  sz = NULL,
  u = NULL,
  vertAlign = NULL,
  scheme = NULL,
  ...
)
}
\arguments{
\item{rFont}{rFont}

\item{charset}{charset}

\item{family}{family}

\item{b}{b}

\item{i}{i}

\item{strike}{strike}

\item{outline}{outline}

\item{shadow}{shadow}

\item{condense}{condense}

\item{extend}{extend}

\item{color}{color}

\item{sz}{sz}

\item{u}{u}

\item{vertAlign}{vertAlign}

\item{scheme}{scheme}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Font for inline text because, yes what you need are different objects with the same elements but different constraints.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.text.html}
}
\examples{
\dontrun{
InlineFont()
}
}
