% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerShadowEffect.R
\name{InnerShadowEffect}
\alias{InnerShadowEffect}
\title{InnerShadowEffect}
\usage{
InnerShadowEffect(blurRad = NULL, dist = NULL, dir = NULL, ...)
}
\arguments{
\item{blurRad}{blurRad}

\item{dist}{dist}

\item{dir}{dir}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
InnerShadowEffect(blurRad = 1, dist = 1, dir = 1)
}
}
