% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputCells.R
\name{InputCells}
\alias{InputCells}
\title{InputCells}
\usage{
InputCells(
  r = NULL,
  deleted = FALSE,
  undone = FALSE,
  val = NULL,
  numFmtId = NULL,
  ...
)
}
\arguments{
\item{r}{r}

\item{deleted}{deleted}

\item{undone}{undone}

\item{val}{val}

\item{numFmtId}{numFmtId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.scenario.html}
}
\examples{
\dontrun{
InputCells(r = "r", val = "val")
}
}
