% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_openpyxl.R
\name{install_openpyxl}
\alias{install_openpyxl}
\title{Install openpyxl Python module}
\usage{
install_openpyxl(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method}

\item{conda}{conda}
}
\value{
(Invisible) `TRUE` if the Conda environment was created successfully.
}
\description{
Install openpyxl Python module
}
