% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pillow.R
\name{install_pillow}
\alias{install_pillow}
\title{Install pillow Python module}
\usage{
install_pillow(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method}

\item{conda}{conda}
}
\value{
(Invisible) `TRUE` if the Conda environment was created successfully.
}
\description{
Module needed to include images in an Excel file.
}
