% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{Layout}
\alias{Layout}
\title{Layout}
\usage{
Layout(manualLayout = NULL, extLst = NULL, ...)
}
\arguments{
\item{manualLayout}{manualLayout}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.layout.html}
}
\examples{
\dontrun{
Layout()
}
}
