% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Legend.R
\name{Legend}
\alias{Legend}
\title{Legend}
\usage{
Legend(
  legendPos = "r",
  legendEntry = list(),
  layout = NULL,
  overlay = NULL,
  spPr = NULL,
  txPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{legendPos}{legendPos}

\item{legendEntry}{legendEntry}

\item{layout}{layout}

\item{overlay}{overlay}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.legend.html}
}
\examples{
\dontrun{
Legend()
}
}
