% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegendEntry.R
\name{LegendEntry}
\alias{LegendEntry}
\title{LegendEntry}
\usage{
LegendEntry(idx = 0, delete = NULL, txPr = NULL, extLst = NULL, ...)
}
\arguments{
\item{idx}{idx}

\item{delete}{delete}

\item{txPr}{txPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.legend.html}
}
\examples{
\dontrun{
LegendEntry()
}
}
