% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Level.R
\name{Level}
\alias{Level}
\title{Level}
\usage{
Level(pt = list(), ...)
}
\arguments{
\item{pt}{pt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
Level()
}
}
