% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightRig.R
\name{LightRig}
\alias{LightRig}
\title{LightRig}
\usage{
LightRig(rig = NULL, dir = NULL, rot = NULL, ...)
}
\arguments{
\item{rig}{rig}

\item{dir}{dir}

\item{rot}{rot}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
LightRig(rig = "morning", dir = "t")
}
}
