% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearShadeProperties.R
\name{LinearShadeProperties}
\alias{LinearShadeProperties}
\title{LinearShadeProperties}
\usage{
LinearShadeProperties(ang = NULL, scaled = NULL, ...)
}
\arguments{
\item{ang}{ang}

\item{scaled}{scaled}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
LinearShadeProperties(ang = 30)
}
}
