% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineBreak.R
\name{LineBreak}
\alias{LineBreak}
\title{LineBreak}
\usage{
LineBreak(rPr = NULL, ...)
}
\arguments{
\item{rPr}{rPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
LineBreak()
}
}
