% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart.R
\name{LineChart}
\alias{LineChart}
\title{LineChart}
\usage{
LineChart(
  hiLowLines = NULL,
  upDownBars = NULL,
  marker = NULL,
  smooth = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{hiLowLines}{hiLowLines}

\item{upDownBars}{upDownBars}

\item{marker}{marker}

\item{smooth}{smooth}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.line_chart.html}
}
\examples{
\dontrun{
LineChart()
}
}
