% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart3D.R
\name{LineChart3D}
\alias{LineChart3D}
\title{LineChart3D}
\usage{
LineChart3D(
  gapDepth = NULL,
  hiLowLines = NULL,
  upDownBars = NULL,
  marker = NULL,
  smooth = NULL,
  ...
)
}
\arguments{
\item{gapDepth}{gapDepth}

\item{hiLowLines}{hiLowLines}

\item{upDownBars}{upDownBars}

\item{marker}{marker}

\item{smooth}{smooth}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.line_chart.html}
}
\examples{
\dontrun{
LineChart3D()
}
}
