% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineEndProperties.R
\name{LineEndProperties}
\alias{LineEndProperties}
\title{LineEndProperties}
\usage{
LineEndProperties(type = NULL, w = NULL, len = NULL, ...)
}
\arguments{
\item{type}{type}

\item{w}{w}

\item{len}{len}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.line.html}
}
\examples{
\dontrun{
LineEndProperties()
}
}
