% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineProperties.R
\name{LineProperties}
\alias{LineProperties}
\title{LineProperties}
\usage{
LineProperties(
  w = NULL,
  cap = NULL,
  cmpd = NULL,
  algn = NULL,
  noFill = NULL,
  solidFill = NULL,
  gradFill = NULL,
  pattFill = NULL,
  prstDash = NULL,
  custDash = NULL,
  round = NULL,
  bevel = NULL,
  miter = NULL,
  headEnd = NULL,
  tailEnd = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{w}{w}

\item{cap}{cap}

\item{cmpd}{cmpd}

\item{algn}{algn}

\item{noFill}{noFill}

\item{solidFill}{solidFill}

\item{gradFill}{gradFill}

\item{pattFill}{pattFill}

\item{prstDash}{prstDash}

\item{custDash}{custDash}

\item{round}{round}

\item{bevel}{bevel}

\item{miter}{miter}

\item{headEnd}{headEnd}

\item{tailEnd}{tailEnd}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.line.html}
}
\examples{
\dontrun{
LineProperties()
}
}
