% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListStyle.R
\name{ListStyle}
\alias{ListStyle}
\title{ListStyle}
\usage{
ListStyle(
  defPPr = NULL,
  lvl1pPr = NULL,
  lvl2pPr = NULL,
  lvl3pPr = NULL,
  lvl4pPr = NULL,
  lvl5pPr = NULL,
  lvl6pPr = NULL,
  lvl7pPr = NULL,
  lvl8pPr = NULL,
  lvl9pPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{defPPr}{defPPr}

\item{lvl1pPr}{lvl1pPr}

\item{lvl2pPr}{lvl2pPr}

\item{lvl3pPr}{lvl3pPr}

\item{lvl4pPr}{lvl4pPr}

\item{lvl5pPr}{lvl5pPr}

\item{lvl6pPr}{lvl6pPr}

\item{lvl7pPr}{lvl7pPr}

\item{lvl8pPr}{lvl8pPr}

\item{lvl9pPr}{lvl9pPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
ListStyle()
}
}
