% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuminanceEffect.R
\name{LuminanceEffect}
\alias{LuminanceEffect}
\title{LuminanceEffect}
\usage{
LuminanceEffect(bright = 0L, contrast = 0L, ...)
}
\arguments{
\item{bright}{bright}

\item{contrast}{contrast}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
LuminanceEffect()
}

}
