% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualLayout.R
\name{ManualLayout}
\alias{ManualLayout}
\title{ManualLayout}
\usage{
ManualLayout(
  layoutTarget = NULL,
  xMode = NULL,
  yMode = NULL,
  wMode = "factor",
  hMode = "factor",
  x = NULL,
  y = NULL,
  w = NULL,
  h = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{layoutTarget}{layoutTarget}

\item{xMode}{xMode}

\item{yMode}{yMode}

\item{wMode}{wMode}

\item{hMode}{hMode}

\item{x}{x}

\item{y}{y}

\item{w}{w}

\item{h}{h}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.layout.html}
}
\examples{
\dontrun{
ManualLayout()
}

}
