% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker.R
\name{Marker}
\alias{Marker}
\title{Marker}
\usage{
Marker(symbol = NULL, size = NULL, spPr = NULL, extLst = NULL, ...)
}
\arguments{
\item{symbol}{symbol}

\item{size}{size}

\item{spPr}{spPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.marker.html}
}
\examples{
\dontrun{
Marker()
}

}
