% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeCell.R
\name{MergeCell}
\alias{MergeCell}
\title{MergeCell}
\usage{
MergeCell(ref = NULL, ...)
}
\arguments{
\item{ref}{ref}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Represents a range in a sheet: title and coordinates.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.merge.html}
}
\details{
This object is used to perform operations on ranges, like: - shift, expand or shrink
- union/intersection with another sheet range, We can check whether a range is: - equal or not equal to another,
- disjoint of another,
- contained in another. We can get: - the size of a range.
- the range bounds (vertices)
- the coordinates,
- the string representation,
}
\examples{
\dontrun{
MergeCell(ref = "A1")
}

}
