% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeCells.R
\name{MergeCells}
\alias{MergeCells}
\title{MergeCells}
\usage{
MergeCells(count = NULL, mergeCell = list(), ...)
}
\arguments{
\item{count}{count}

\item{mergeCell}{mergeCell}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.merge.html}
}
\examples{
\dontrun{
MergeCells()
}

}
