% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergedCell.R
\name{MergedCell}
\alias{MergedCell}
\title{MergedCell}
\usage{
MergedCell(worksheet, row = NULL, column = NULL, ...)
}
\arguments{
\item{worksheet}{worksheet}

\item{row}{row}

\item{column}{column}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Describes the properties of a cell in a merged cell and helps to display the borders of the merged cell.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.cell.html}
}
\details{
The value of a MergedCell is always NULL.
}
\examples{
\dontrun{
wb <- Workbook()
MergedCell(worksheet = wb$active, row = 1L, column = 1L)
}

}
