% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCellRange.R
\name{MultiCellRange}
\alias{MultiCellRange}
\title{MultiCellRange}
\usage{
MultiCellRange(ranges = list(), ...)
}
\arguments{
\item{ranges}{ranges}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.cell_range.html}
}
\examples{
\dontrun{
MultiCellRange()
}

}
