% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLevelStrData.R
\name{MultiLevelStrData}
\alias{MultiLevelStrData}
\title{MultiLevelStrData}
\usage{
MultiLevelStrData(ptCount = NULL, lvl = list(), extLst = NULL, ...)
}
\arguments{
\item{ptCount}{ptCount}

\item{lvl}{lvl}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
MultiLevelStrData()
}

}
