% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLevelStrRef.R
\name{MultiLevelStrRef}
\alias{MultiLevelStrRef}
\title{MultiLevelStrRef}
\usage{
MultiLevelStrRef(f = NULL, multiLvlStrCache = NULL, extLst = NULL, ...)
}
\arguments{
\item{f}{f}

\item{multiLvlStrCache}{multiLvlStrCache}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
MultiLevelStrRef()
}

}
