% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedStyle.R
\name{NamedStyle}
\alias{NamedStyle}
\title{NamedStyle}
\usage{
NamedStyle(
  name = "Normal",
  font = NULL,
  fill = NULL,
  border = NULL,
  alignment = NULL,
  number_format = NULL,
  protection = NULL,
  builtinId = NULL,
  hidden = FALSE,
  xfId = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{font}{font}

\item{fill}{fill}

\item{border}{border}

\item{alignment}{alignment}

\item{number_format}{number_format}

\item{protection}{protection}

\item{builtinId}{builtinId}

\item{hidden}{hidden}

\item{xfId}{xfId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Named and editable styles

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.named_styles.html}
}
\examples{
\dontrun{
NamedStyle()
}

}
