% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedStyleList.R
\name{NamedStyleList}
\alias{NamedStyleList}
\title{NamedStyleList}
\usage{
NamedStyleList(iterable = list(), ...)
}
\arguments{
\item{iterable}{iterable}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Named styles are editable and can be applied to multiple objects

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.named_styles.html}
}
\details{
As only the index is stored in referencing objects the order mus
be preserved.
}
\examples{
\dontrun{
NamedStyleList()
}

}
