% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedOverlap.R
\name{NestedOverlap}
\alias{NestedOverlap}
\title{NestedOverlap}
\usage{
NestedOverlap(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.descriptors.html}
}
\examples{
\dontrun{
NestedOverlap()
}

}
