% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonVisualConnectorProperties.R
\name{NonVisualConnectorProperties}
\alias{NonVisualConnectorProperties}
\title{NonVisualConnectorProperties}
\usage{
NonVisualConnectorProperties(
  cxnSpLocks = NULL,
  stCxn = NULL,
  endCxn = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{cxnSpLocks}{cxnSpLocks}

\item{stCxn}{stCxn}

\item{endCxn}{endCxn}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.connector.html}
}
\examples{
\dontrun{
NonVisualConnectorProperties()
}

}
