% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonVisualDrawingShapeProps.R
\name{NonVisualDrawingShapeProps}
\alias{NonVisualDrawingShapeProps}
\title{NonVisualDrawingShapeProps}
\usage{
NonVisualDrawingShapeProps(spLocks = NULL, txBox = NULL, extLst = NULL, ...)
}
\arguments{
\item{spLocks}{spLocks}

\item{txBox}{txBox}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.properties.html}
}
\examples{
\dontrun{
NonVisualDrawingShapeProps()
}

}
