% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonVisualGraphicFrameProperties.R
\name{NonVisualGraphicFrameProperties}
\alias{NonVisualGraphicFrameProperties}
\title{NonVisualGraphicFrameProperties}
\usage{
NonVisualGraphicFrameProperties(graphicFrameLocks = NULL, extLst = NULL, ...)
}
\arguments{
\item{graphicFrameLocks}{graphicFrameLocks}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.graphic.html}
}
\examples{
\dontrun{
NonVisualGraphicFrameProperties()
}

}
