% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonVisualPictureProperties.R
\name{NonVisualPictureProperties}
\alias{NonVisualPictureProperties}
\title{NonVisualPictureProperties}
\usage{
NonVisualPictureProperties(
  preferRelativeResize = NULL,
  picLocks = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{preferRelativeResize}{preferRelativeResize}

\item{picLocks}{picLocks}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.picture.html}
}
\examples{
\dontrun{
NonVisualPictureProperties()
}

}
