% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataSource.R
\name{NumDataSource}
\alias{NumDataSource}
\title{NumDataSource}
\usage{
NumDataSource(numRef = NULL, numLit = NULL, ...)
}
\arguments{
\item{numRef}{numRef}

\item{numLit}{numLit}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
NumDataSource()
}

}
