% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericAxis.R
\name{NumericAxis}
\alias{NumericAxis}
\title{NumericAxis}
\usage{
NumericAxis(
  crossBetween = NULL,
  majorUnit = NULL,
  minorUnit = NULL,
  dispUnits = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{crossBetween}{crossBetween}

\item{majorUnit}{majorUnit}

\item{minorUnit}{minorUnit}

\item{dispUnits}{dispUnits}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
NumericAxis()
}

}
