% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumFmt.R
\name{NumFmt}
\alias{NumFmt}
\title{NumFmt}
\usage{
NumFmt(formatCode = NULL, sourceLinked = FALSE, ...)
}
\arguments{
\item{formatCode}{formatCode}

\item{sourceLinked}{sourceLinked}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
NumFmt(formatCode = "formatCode")
}

}
