% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRef.R
\name{NumRef}
\alias{NumRef}
\title{NumRef}
\usage{
NumRef(f = NULL, numCache = NULL, extLst = NULL, ...)
}
\arguments{
\item{f}{f}

\item{numCache}{numCache}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
NumRef()
}

}
