% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumVal.R
\name{NumVal}
\alias{NumVal}
\title{NumVal}
\usage{
NumVal(idx = NULL, formatCode = NULL, v = NULL)
}
\arguments{
\item{idx}{idx}

\item{formatCode}{formatCode}

\item{v}{v}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
NumVal(idx = 1L)
}

}
