% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectAnchor.R
\name{ObjectAnchor}
\alias{ObjectAnchor}
\title{ObjectAnchor}
\usage{
ObjectAnchor(
  from = NULL,
  to = NULL,
  moveWithCells = FALSE,
  sizeWithCells = FALSE,
  z_order = NULL,
  ...
)
}
\arguments{
\item{from}{_from}

\item{to}{to}

\item{moveWithCells}{moveWithCells}

\item{sizeWithCells}{sizeWithCells}

\item{z_order}{z_order}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.ole.html}
}
\examples{
\dontrun{
ObjectAnchor(from = AnchorMarker(), to = AnchorMarker())
}

}
