% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneCellAnchor.R
\name{OneCellAnchor}
\alias{OneCellAnchor}
\title{OneCellAnchor}
\usage{
OneCellAnchor(from = NULL, ext = NULL, ...)
}
\arguments{
\item{from}{from}

\item{ext}{ext}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Note that the argument `_from` has been transformed to `from`.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
OneCellAnchor()
}

}
