% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outline.R
\name{Outline}
\alias{Outline}
\title{Outline}
\usage{
Outline(
  applyStyles = NULL,
  summaryBelow = NULL,
  summaryRight = NULL,
  showOutlineSymbols = NULL,
  ...
)
}
\arguments{
\item{applyStyles}{applyStyles}

\item{summaryBelow}{summaryBelow}

\item{summaryRight}{summaryRight}

\item{showOutlineSymbols}{showOutlineSymbols}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.properties.html}
}
\examples{
\dontrun{
Outline()
}

}
