% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageBreak.R
\name{PageBreak}
\alias{PageBreak}
\title{PageBreak}
\usage{
PageBreak(count = NULL, manualBreakCount = NULL, brk = list(), ...)
}
\arguments{
\item{count}{count}

\item{manualBreakCount}{manualBreakCount}

\item{brk}{brk}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.pagebreak.html}
}
\examples{
\dontrun{
PageBreak()
}

}
