% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageMargins.R
\name{PageMargins}
\alias{PageMargins}
\title{PageMargins}
\usage{
PageMargins(
  left = 0.75,
  right = 0.75,
  top = 1L,
  bottom = 1L,
  header = 0.5,
  footer = 0.5,
  ...
)
}
\arguments{
\item{left}{left}

\item{right}{right}

\item{top}{top}

\item{bottom}{bottom}

\item{header}{header}

\item{footer}{footer}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Information about page margins for view/print layouts.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.page.html}
}
\details{
Standard values (in inches)
left, right = 0.75
top, bottom = 1
header, footer = 0.5
}
\examples{
\dontrun{
PageMargins()
}

}
