% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageSetupProperties.R
\name{PageSetupProperties}
\alias{PageSetupProperties}
\title{PageSetupProperties}
\usage{
PageSetupProperties(autoPageBreaks = NULL, fitToPage = NULL, ...)
}
\arguments{
\item{autoPageBreaks}{autoPageBreaks}

\item{fitToPage}{fitToPage}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.properties.html}
}
\examples{
\dontrun{
PageSetupProperties()
}

}
