% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pane.R
\name{Pane}
\alias{Pane}
\title{Pane}
\usage{
Pane(
  xSplit = NULL,
  ySplit = NULL,
  topLeftCell = NULL,
  activePane = "topLeft",
  state = "split",
  ...
)
}
\arguments{
\item{xSplit}{xSplit}

\item{ySplit}{ySplit}

\item{topLeftCell}{topLeftCell}

\item{activePane}{activePane}

\item{state}{state}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.views.html}
}
\examples{
\dontrun{
Pane()
}

}
