% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paragraph.R
\name{Paragraph}
\alias{Paragraph}
\title{Paragraph}
\usage{
Paragraph(pPr = NULL, endParaRPr = NULL, r = NULL, br = NULL, fld = NULL, ...)
}
\arguments{
\item{pPr}{pPr}

\item{endParaRPr}{endParaRPr}

\item{r}{r}

\item{br}{br}

\item{fld}{fld}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
Paragraph()
}

}
