% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParagraphProperties.R
\name{ParagraphProperties}
\alias{ParagraphProperties}
\title{ParagraphProperties}
\usage{
ParagraphProperties(
  marL = NULL,
  marR = NULL,
  lvl = NULL,
  indent = NULL,
  algn = NULL,
  defTabSz = NULL,
  rtl = NULL,
  eaLnBrk = NULL,
  fontAlgn = NULL,
  latinLnBrk = NULL,
  hangingPunct = NULL,
  lnSpc = NULL,
  spcBef = NULL,
  spcAft = NULL,
  tabLst = NULL,
  defRPr = NULL,
  extLst = NULL,
  buClrTx = NULL,
  buClr = NULL,
  buSzTx = NULL,
  buSzPct = NULL,
  buSzPts = NULL,
  buFontTx = NULL,
  buFont = NULL,
  buNone = NULL,
  buAutoNum = NULL,
  buChar = NULL,
  buBlip = NULL,
  ...
)
}
\arguments{
\item{marL}{marL}

\item{marR}{marR}

\item{lvl}{lvl}

\item{indent}{indent}

\item{algn}{algn}

\item{defTabSz}{defTabSz}

\item{rtl}{rtl}

\item{eaLnBrk}{eaLnBrk}

\item{fontAlgn}{fontAlgn}

\item{latinLnBrk}{latinLnBrk}

\item{hangingPunct}{hangingPunct}

\item{lnSpc}{lnSpc}

\item{spcBef}{spcBef}

\item{spcAft}{spcAft}

\item{tabLst}{tabLst}

\item{defRPr}{defRPr}

\item{extLst}{extLst}

\item{buClrTx}{buClrTx}

\item{buClr}{buClr}

\item{buSzTx}{buSzTx}

\item{buSzPct}{buSzPct}

\item{buSzPts}{buSzPts}

\item{buFontTx}{buFontTx}

\item{buFont}{buFont}

\item{buNone}{buNone}

\item{buAutoNum}{buAutoNum}

\item{buChar}{buChar}

\item{buBlip}{buBlip}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
ParagraphProperties()
}

}
