% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path2D.R
\name{Path2D}
\alias{Path2D}
\title{Path2D}
\usage{
Path2D(w = NULL, h = NULL, fill = NULL, stroke = NULL, extrusionOk = NULL, ...)
}
\arguments{
\item{w}{w}

\item{h}{h}

\item{fill}{fill}

\item{stroke}{stroke}

\item{extrusionOk}{extrusionOk}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Path2D(w = 1L, h = 1L)
}

}
