% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path2DList.R
\name{Path2DList}
\alias{Path2DList}
\title{Path2DList}
\usage{
Path2DList(path = NULL, ...)
}
\arguments{
\item{path}{path}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Path2DList()
}

}
