% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathShadeProperties.R
\name{PathShadeProperties}
\alias{PathShadeProperties}
\title{PathShadeProperties}
\usage{
PathShadeProperties(path = NULL, fillToRect = NULL, ...)
}
\arguments{
\item{path}{path}

\item{fillToRect}{fillToRect}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
PathShadeProperties(path = "circle")
}

}
