% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatternFill.R
\name{PatternFill}
\alias{PatternFill}
\title{PatternFill}
\usage{
PatternFill(
  patternType = NULL,
  fill_type = NULL,
  start_color = NULL,
  end_color = NULL,
  ...
)
}
\arguments{
\item{patternType}{patternType}

\item{fill_type}{fill_type}

\item{start_color}{start_color}

\item{end_color}{end_color}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Area fill patterns for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.fills.html}
}
\details{
Caution: if you do not specify a fill_type, other attributes will have
no effect !
}
\section{Caution}{
 if you do not specify a fill_type, other attributes will have:
no effect !
}

\examples{
\dontrun{
PatternFill()
}

}
