% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatternFillProperties.R
\name{PatternFillProperties}
\alias{PatternFillProperties}
\title{PatternFillProperties}
\usage{
PatternFillProperties(prst = NULL, fgClr = NULL, bgClr = NULL, ...)
}
\arguments{
\item{prst}{prst}

\item{fgClr}{fgClr}

\item{bgClr}{bgClr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
PatternFillProperties()
}

}
