% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhoneticProperties.R
\name{PhoneticProperties}
\alias{PhoneticProperties}
\title{PhoneticProperties}
\usage{
PhoneticProperties(fontId = NULL, type = NULL, alignment = NULL, ...)
}
\arguments{
\item{fontId}{fontId}

\item{type}{type}

\item{alignment}{alignment}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.text.html}
}
\examples{
\dontrun{
PhoneticProperties(fontId = 1L)
}

}
