% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhoneticText.R
\name{PhoneticText}
\alias{PhoneticText}
\title{PhoneticText}
\usage{
PhoneticText(sb = NULL, eb = NULL, t = NULL, ...)
}
\arguments{
\item{sb}{sb}

\item{eb}{eb}

\item{t}{t}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.text.html}
}
\examples{
\dontrun{
PhoneticText(sb = 1L, eb = 1L)
}

}
