% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart.R
\name{PieChart}
\alias{PieChart}
\title{PieChart}
\usage{
PieChart(firstSliceAng = 0L, extLst = NULL, ...)
}
\arguments{
\item{firstSliceAng}{firstSliceAng}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pie_chart.html}
}
\examples{
\dontrun{
PieChart()
}

}
