% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart3D.R
\name{PieChart3D}
\alias{PieChart3D}
\title{PieChart3D}
\usage{
PieChart3D(varyColors = TRUE, ser = list(), dLbls = NULL, ...)
}
\arguments{
\item{varyColors}{varyColors}

\item{ser}{ser}

\item{dLbls}{dLbls}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pie_chart.html}
}
\examples{
\dontrun{
PieChart3D()
}

}
