% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotFormat.R
\name{PivotFormat}
\alias{PivotFormat}
\title{PivotFormat}
\usage{
PivotFormat(
  idx = 0L,
  spPr = NULL,
  txPr = NULL,
  marker = NULL,
  dLbl = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{idx}{idx}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{marker}{marker}

\item{dLbl}{dLbl}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pivot.html}
}
\examples{
\dontrun{
PivotFormat()
}

}
