% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotSource.R
\name{PivotSource}
\alias{PivotSource}
\title{PivotSource}
\usage{
PivotSource(name = NULL, fmtId = NULL, extLst = NULL, ...)
}
\arguments{
\item{name}{name}

\item{fmtId}{fmtId}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pivot.html}
}
\examples{
\dontrun{
PivotSource(fmtId = 1L)
}

}
