% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels_to_EMU.R
\name{pixels_to_EMU}
\alias{pixels_to_EMU}
\title{pixels_to_EMU}
\usage{
pixels_to_EMU(value, ...)
}
\arguments{
\item{value}{value}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
1 pixel = 9525 EMUs

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
pixels_to_EMU(value = 1L)
}

}
