% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels_to_points.R
\name{pixels_to_points}
\alias{pixels_to_points}
\title{pixels_to_points}
\usage{
pixels_to_points(value, dpi = 96L, ...)
}
\arguments{
\item{value}{value}

\item{dpi}{dpi}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
96 dpi, 72i

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
pixels_to_points(value = 1L)
}

}
