% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotArea.R
\name{PlotArea}
\alias{PlotArea}
\title{PlotArea}
\usage{
PlotArea(
  layout = NULL,
  dTable = NULL,
  spPr = NULL,
  charts = list(),
  axes = list(),
  extLst = NULL,
  ...
)
}
\arguments{
\item{layout}{layout}

\item{dTable}{dTable}

\item{spPr}{spPr}

\item{charts}{_charts}

\item{axes}{_axes}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.plotarea.html}
}
\examples{
\dontrun{
PlotArea()
}

}
