% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point3D.R
\name{Point3D}
\alias{Point3D}
\title{Point3D}
\usage{
Point3D(x = NULL, y = NULL, z = NULL, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{z}{z}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Point3D(x = 1L, y = 1L, z = 1L)
}

}
