% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_pixels.R
\name{points_to_pixels}
\alias{points_to_pixels}
\title{points_to_pixels}
\usage{
points_to_pixels(value, dpi = 96L, ...)
}
\arguments{
\item{value}{value}

\item{dpi}{dpi}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
A numeric value.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
points_to_pixels(value = 1L)
}

}
