% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositiveSize2D.R
\name{PositiveSize2D}
\alias{PositiveSize2D}
\title{PositiveSize2D}
\usage{
PositiveSize2D(cx = NULL, cy = NULL, ...)
}
\arguments{
\item{cx}{cx}

\item{cy}{cy}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
PositiveSize2D(cx = 1L, cy = 1L)
}

}
