% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresetGeometry2D.R
\name{PresetGeometry2D}
\alias{PresetGeometry2D}
\title{PresetGeometry2D}
\usage{
PresetGeometry2D(prst = NULL, avLst = NULL, ...)
}
\arguments{
\item{prst}{prst}

\item{avLst}{avLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
PresetGeometry2D(prst = "diamond")
}

}
