% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresetShadowEffect.R
\name{PresetShadowEffect}
\alias{PresetShadowEffect}
\title{PresetShadowEffect}
\usage{
PresetShadowEffect(prst = NULL, dist = NULL, dir = NULL, ...)
}
\arguments{
\item{prst}{prst}

\item{dist}{dist}

\item{dir}{dir}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
PresetShadowEffect(prst = "shdw20", dist = 1L, dir = 1L)
}

}
