% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintArea.R
\name{PrintArea}
\alias{PrintArea}
\title{PrintArea}
\usage{
PrintArea(ranges = list(), title = "", ...)
}
\arguments{
\item{ranges}{ranges}

\item{title}{title}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.print_settings.html}
}
\examples{
\dontrun{
PrintArea()
}

}
