% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintOptions.R
\name{PrintOptions}
\alias{PrintOptions}
\title{PrintOptions}
\usage{
PrintOptions(
  horizontalCentered = NULL,
  verticalCentered = NULL,
  headings = NULL,
  gridLines = NULL,
  gridLinesSet = NULL,
  ...
)
}
\arguments{
\item{horizontalCentered}{horizontalCentered}

\item{verticalCentered}{verticalCentered}

\item{headings}{headings}

\item{gridLines}{gridLines}

\item{gridLinesSet}{gridLinesSet}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Worksheet print options

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.page.html}
}
\examples{
\dontrun{
PrintOptions()
}

}
