% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintPageSetup.R
\name{PrintPageSetup}
\alias{PrintPageSetup}
\title{PrintPageSetup}
\usage{
PrintPageSetup(
  worksheet = NULL,
  orientation = NULL,
  paperSize = NULL,
  scale = NULL,
  fitToHeight = NULL,
  fitToWidth = NULL,
  firstPageNumber = NULL,
  useFirstPageNumber = NULL,
  paperHeight = NULL,
  paperWidth = NULL,
  pageOrder = NULL,
  usePrinterDefaults = NULL,
  blackAndWhite = NULL,
  draft = NULL,
  cellComments = NULL,
  errors = NULL,
  horizontalDpi = NULL,
  verticalDpi = NULL,
  copies = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{worksheet}{worksheet}

\item{orientation}{orientation}

\item{paperSize}{paperSize}

\item{scale}{scale}

\item{fitToHeight}{fitToHeight}

\item{fitToWidth}{fitToWidth}

\item{firstPageNumber}{firstPageNumber}

\item{useFirstPageNumber}{useFirstPageNumber}

\item{paperHeight}{paperHeight}

\item{paperWidth}{paperWidth}

\item{pageOrder}{pageOrder}

\item{usePrinterDefaults}{usePrinterDefaults}

\item{blackAndWhite}{blackAndWhite}

\item{draft}{draft}

\item{cellComments}{cellComments}

\item{errors}{errors}

\item{horizontalDpi}{horizontalDpi}

\item{verticalDpi}{verticalDpi}

\item{copies}{copies}

\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Worksheet print page setup

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.page.html}
}
\examples{
\dontrun{
PrintPageSetup()
}

}
