% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSettings.R
\name{PrintSettings}
\alias{PrintSettings}
\title{PrintSettings}
\usage{
PrintSettings(headerFooter = NULL, pageMargins = NULL, pageSetup = NULL, ...)
}
\arguments{
\item{headerFooter}{headerFooter}

\item{pageMargins}{pageMargins}

\item{pageSetup}{pageSetup}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.print_settings.html}
}
\examples{
\dontrun{
PrintSettings()
}

}
