% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTitles.R
\name{PrintTitles}
\alias{PrintTitles}
\title{PrintTitles}
\usage{
PrintTitles(cols = NULL, rows = NULL, title = "", ...)
}
\arguments{
\item{cols}{cols}

\item{rows}{rows}

\item{title}{title}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Contains at least either a range of rows or columns

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.print_settings.html}
}
\examples{
\dontrun{
PrintTitles()
}

}
