% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectedPieChart.R
\name{ProjectedPieChart}
\alias{ProjectedPieChart}
\title{ProjectedPieChart}
\usage{
ProjectedPieChart(
  ofPieType = "pie",
  gapWidth = NULL,
  splitType = "auto",
  splitPos = NULL,
  custSplit = NULL,
  secondPieSize = 75L,
  serLines = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{ofPieType}{ofPieType}

\item{gapWidth}{gapWidth}

\item{splitType}{splitType}

\item{splitPos}{splitPos}

\item{custSplit}{custSplit}

\item{secondPieSize}{secondPieSize}

\item{serLines}{serLines}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.pie_chart.html}
}
\details{
This element contains the pie of pie or bar of pie series on this
chart. Only the first series shall be displayed. The splitType element
shall determine whether the splitPos and custSplit elements apply.
}
\examples{
\dontrun{
ProjectedPieChart()
}

}
