% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protection.R
\name{Protection}
\alias{Protection}
\title{Protection}
\usage{
Protection(locked = TRUE, hidden = FALSE, ...)
}
\arguments{
\item{locked}{locked}

\item{hidden}{hidden}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Protection options for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.protection.html}
}
\examples{
\dontrun{
Protection()
}

}
