% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_sheetname.R
\name{quote_sheetname}
\alias{quote_sheetname}
\title{quote_sheetname}
\usage{
quote_sheetname(sheetname, ...)
}
\arguments{
\item{sheetname}{sheetname}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An string.
}
\description{
Add quotes around sheetnames if they contain spaces.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\examples{
\dontrun{
wb <- Workbook()
quote_sheetname(wb$sheetnames)
}

}
