% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadarChart.R
\name{RadarChart}
\alias{RadarChart}
\title{RadarChart}
\usage{
RadarChart(
  radarStyle = "standard",
  varyColors = NULL,
  ser = list(),
  dLbls = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{radarStyle}{radarStyle}

\item{varyColors}{varyColors}

\item{ser}{ser}

\item{dLbls}{dLbls}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.radar_chart.html}
}
\examples{
\dontrun{
RadarChart()
}

}
