% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_boundaries.R
\name{range_boundaries}
\alias{range_boundaries}
\title{range_boundaries}
\usage{
range_boundaries(range_string, ...)
}
\arguments{
\item{range_string}{range_string}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An list.
}
\description{
Convert a range string into a list of boundaries:

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\details{
(min_col, min_row, max_col, max_row)
Cell coordinates will be converted into a range with the cell at both end
}
\section{Convert a range string into a tuple of boundaries}{

(min_col, min_row, max_col, max_row) Cell coordinates will be converted into a range with the cell at both end
}

\examples{
\dontrun{
range_boundaries(range_string = "A")
}

}
