% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_to_tuple.R
\name{range_to_tuple}
\alias{range_to_tuple}
\title{range_to_tuple}
\usage{
range_to_tuple(range_string, ...)
}
\arguments{
\item{range_string}{range_string}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An list.
}
\description{
Convert a worksheet range to the sheetname and maximum and minimum
coordinate indices

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.cell.html}
}
\examples{
\dontrun{
range_to_tuple(range_string = "sheetname!A1:E4")
}

}
