% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadOnlyCell.R
\name{ReadOnlyCell}
\alias{ReadOnlyCell}
\title{ReadOnlyCell}
\usage{
ReadOnlyCell(sheet, row, column, value, data_type = "n", style_id = 0L, ...)
}
\arguments{
\item{sheet}{sheet}

\item{row}{row}

\item{column}{column}

\item{value}{value}

\item{data_type}{data_type}

\item{style_id}{style_id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.read_only.html}
}
\examples{
\dontrun{
ReadOnlyCell(sheet = "sheet", row = 1L, column = 1L, value = 1L)
}

}
