% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularTextRun.R
\name{RegularTextRun}
\alias{RegularTextRun}
\title{RegularTextRun}
\usage{
RegularTextRun(rPr = NULL, t = "", ...)
}
\arguments{
\item{rPr}{rPr}

\item{t}{t}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
RegularTextRun()
}

}
