% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Related.R
\name{Related}
\alias{Related}
\title{Related}
\usage{
Related(id = NULL, ...)
}
\arguments{
\item{id}{id}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.related.html}
}
\examples{
\dontrun{
Related()
}

}
