% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRect.R
\name{RelativeRect}
\alias{RelativeRect}
\title{RelativeRect}
\usage{
RelativeRect(l = NULL, t = NULL, r = NULL, b = NULL, ...)
}
\arguments{
\item{l}{l}

\item{t}{t}

\item{r}{r}

\item{b}{b}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
RelativeRect()
}

}
