% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RichTextProperties.R
\name{RichTextProperties}
\alias{RichTextProperties}
\title{RichTextProperties}
\usage{
RichTextProperties(
  rot = NULL,
  spcFirstLastPara = NULL,
  vertOverflow = NULL,
  horzOverflow = NULL,
  vert = NULL,
  wrap = NULL,
  lIns = NULL,
  tIns = NULL,
  rIns = NULL,
  bIns = NULL,
  numCol = NULL,
  spcCol = NULL,
  rtlCol = NULL,
  fromWordArt = NULL,
  anchor = NULL,
  anchorCtr = NULL,
  forceAA = NULL,
  upright = NULL,
  compatLnSpc = NULL,
  prstTxWarp = NULL,
  scene3d = NULL,
  extLst = NULL,
  noAutofit = NULL,
  normAutofit = NULL,
  spAutoFit = NULL,
  flatTx = NULL,
  ...
)
}
\arguments{
\item{rot}{rot}

\item{spcFirstLastPara}{spcFirstLastPara}

\item{vertOverflow}{vertOverflow}

\item{horzOverflow}{horzOverflow}

\item{vert}{vert}

\item{wrap}{wrap}

\item{lIns}{lIns}

\item{tIns}{tIns}

\item{rIns}{rIns}

\item{bIns}{bIns}

\item{numCol}{numCol}

\item{spcCol}{spcCol}

\item{rtlCol}{rtlCol}

\item{fromWordArt}{fromWordArt}

\item{anchor}{anchor}

\item{anchorCtr}{anchorCtr}

\item{forceAA}{forceAA}

\item{upright}{upright}

\item{compatLnSpc}{compatLnSpc}

\item{prstTxWarp}{prstTxWarp}

\item{scene3d}{scene3d}

\item{extLst}{extLst}

\item{noAutofit}{noAutofit}

\item{normAutofit}{normAutofit}

\item{spAutoFit}{spAutoFit}

\item{flatTx}{flatTx}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
RichTextProperties()
}

}
