% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowRange.R
\name{RowRange}
\alias{RowRange}
\title{RowRange}
\usage{
RowRange(range_string = NULL, min_row = NULL, max_row = NULL, ...)
}
\arguments{
\item{range_string}{range_string}

\item{min_row}{min_row}

\item{max_row}{max_row}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Represent a range of at least one row

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.print_settings.html}
}
\examples{
\dontrun{
RowRange(min_row = 1L, max_row = 2L)
}

}
