% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule.R
\name{Rule}
\alias{Rule}
\title{Rule}
\usage{
Rule(
  type,
  dxfId = NULL,
  priority = 0L,
  stopIfTrue = NULL,
  aboveAverage = NULL,
  percent = NULL,
  bottom = NULL,
  operator = NULL,
  text = NULL,
  timePeriod = NULL,
  rank = NULL,
  stdDev = NULL,
  equalAverage = NULL,
  formula = list(),
  colorScale = NULL,
  dataBar = NULL,
  iconSet = NULL,
  extLst = NULL,
  dxf = NULL,
  ...
)
}
\arguments{
\item{type}{type}

\item{dxfId}{dxfId}

\item{priority}{priority}

\item{stopIfTrue}{stopIfTrue}

\item{aboveAverage}{aboveAverage}

\item{percent}{percent}

\item{bottom}{bottom}

\item{operator}{operator}

\item{text}{text}

\item{timePeriod}{timePeriod}

\item{rank}{rank}

\item{stdDev}{stdDev}

\item{equalAverage}{equalAverage}

\item{formula}{formula}

\item{colorScale}{colorScale}

\item{dataBar}{dataBar}

\item{iconSet}{iconSet}

\item{extLst}{extLst}

\item{dxf}{dxf}

\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
Rule(type = "colorScale")
}

}
