% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleType.R
\name{RuleType}
\alias{RuleType}
\title{RuleType}
\usage{
RuleType(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}
}
\examples{
\dontrun{
RuleType()
}

}
