% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_workbook.R
\name{save_workbook}
\alias{save_workbook}
\title{save_workbook}
\usage{
save_workbook(workbook, filename, ...)
}
\arguments{
\item{workbook}{workbook}

\item{filename}{filename}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Save the given workbook on the filesystem under the name filename.

\url{https://openpyxl.readthedocs.io/en/latest/api/openpyxl.writer.excel.html}
}
\details{
:param workbook: the workbook to save
:type workbook: :class:`openpyxl.workbook.Workbook` :param filename: the path to which save the workbook
:type filename: string :rtype: bool
}
\examples{
\dontrun{
save_workbook(workbook = wb, filename = tempfile())
}

}
