% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaling.R
\name{Scaling}
\alias{Scaling}
\title{Scaling}
\usage{
Scaling(
  logBase = NULL,
  orientation = "minMax",
  max = NULL,
  min = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{logBase}{logBase}

\item{orientation}{orientation}

\item{max}{max}

\item{min}{min}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
Scaling()
}

}
