% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterChart.R
\name{ScatterChart}
\alias{ScatterChart}
\title{ScatterChart}
\usage{
ScatterChart(
  scatterStyle = NULL,
  varyColors = NULL,
  ser = list(),
  dLbls = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{scatterStyle}{scatterStyle}

\item{varyColors}{varyColors}

\item{ser}{ser}

\item{dLbls}{dLbls}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.scatter_chart.html}
}
\examples{
\dontrun{
ScatterChart()
}

}
