% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenario.R
\name{Scenario}
\alias{Scenario}
\title{Scenario}
\usage{
Scenario(
  inputCells = list(),
  name = NULL,
  locked = FALSE,
  hidden = FALSE,
  count = NULL,
  user = NULL,
  comment = NULL,
  ...
)
}
\arguments{
\item{inputCells}{inputCells}

\item{name}{name}

\item{locked}{locked}

\item{hidden}{hidden}

\item{count}{count}

\item{user}{user}

\item{comment}{comment}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.scenario.html}
}
\examples{
\dontrun{
Scenario(name = "name")
}

}
