% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioList.R
\name{ScenarioList}
\alias{ScenarioList}
\title{ScenarioList}
\usage{
ScenarioList(scenario = list(), current = NULL, show = NULL, sqref = NULL, ...)
}
\arguments{
\item{scenario}{scenario}

\item{current}{current}

\item{show}{show}

\item{sqref}{sqref}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.scenario.html}
}
\examples{
\dontrun{
ScenarioList()
}

}
