% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scene3D.R
\name{Scene3D}
\alias{Scene3D}
\title{Scene3D}
\usage{
Scene3D(camera = NULL, lightRig = NULL, backdrop = NULL, extLst = NULL, ...)
}
\arguments{
\item{camera}{camera}

\item{lightRig}{lightRig}

\item{backdrop}{backdrop}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Scene3D(
camera = Camera(prst = "perspectiveLeft"),
lightRig = LightRig(rig = "morning", dir = "t")
)
}

}
