% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchemeColor.R
\name{SchemeColor}
\alias{SchemeColor}
\title{SchemeColor}
\usage{
SchemeColor(
  tint = NULL,
  shade = NULL,
  comp = NULL,
  inv = NULL,
  gray = NULL,
  alpha = NULL,
  alphaOff = NULL,
  alphaMod = NULL,
  hue = NULL,
  hueOff = NULL,
  hueMod = NULL,
  sat = NULL,
  satOff = NULL,
  satMod = NULL,
  lum = NULL,
  lumOff = NULL,
  lumMod = NULL,
  red = NULL,
  redOff = NULL,
  redMod = NULL,
  green = NULL,
  greenOff = NULL,
  greenMod = NULL,
  blue = NULL,
  blueOff = NULL,
  blueMod = NULL,
  gamma = NULL,
  invGamma = NULL,
  val = NULL,
  ...
)
}
\arguments{
\item{tint}{tint}

\item{shade}{shade}

\item{comp}{comp}

\item{inv}{inv}

\item{gray}{gray}

\item{alpha}{alpha}

\item{alphaOff}{alphaOff}

\item{alphaMod}{alphaMod}

\item{hue}{hue}

\item{hueOff}{hueOff}

\item{hueMod}{hueMod}

\item{sat}{sat}

\item{satOff}{satOff}

\item{satMod}{satMod}

\item{lum}{lum}

\item{lumOff}{lumOff}

\item{lumMod}{lumMod}

\item{red}{red}

\item{redOff}{redOff}

\item{redMod}{redMod}

\item{green}{green}

\item{greenOff}{greenOff}

\item{greenMod}{greenMod}

\item{blue}{blue}

\item{blueOff}{blueOff}

\item{blueMod}{blueMod}

\item{gamma}{gamma}

\item{invGamma}{invGamma}

\item{val}{val}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.colors.html}
}
\examples{
\dontrun{
# from Gradient Pie Charts example
slice <- DataPoint(
 idx = 0,
 explosion = 20,
 spPr = GraphicalProperties(
     gradFill = GradientFillProperties(
         gsLst = list(
                 pos=0,
                 prstClr='blue'
             ),
             GradientStop(
                 pos=100000,
                 schemeClr=SchemeColor(
                     val='accent1',
                     lumMod=30000,
                     lumOff=70000
                 )
             )
         )
     )
 )
}
}
