% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection.R
\name{Selection}
\alias{Selection}
\title{Selection}
\usage{
Selection(
  pane = NULL,
  activeCell = "A1",
  activeCellId = NULL,
  sqref = "A1",
  ...
)
}
\arguments{
\item{pane}{pane}

\item{activeCell}{activeCell}

\item{activeCellId}{activeCellId}

\item{sqref}{sqref}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.views.html}
}
\examples{
\dontrun{
Selection()
}
}
