% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Series.R
\name{Series}
\alias{Series}
\title{Series}
\usage{
Series(
  values,
  xvalues = NULL,
  zvalues = NULL,
  title = NULL,
  title_from_data = FALSE,
  ...
)
}
\arguments{
\item{values}{values}

\item{xvalues}{xvalues}

\item{zvalues}{zvalues}

\item{title}{title}

\item{title_from_data}{title_from_data}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.series.html}
}
\examples{
\dontrun{
Series(values = "sheetname!A1:E4")
}
}
