% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriesAxis.R
\name{SeriesAxis}
\alias{SeriesAxis}
\title{SeriesAxis}
\usage{
SeriesAxis(tickLblSkip = NULL, tickMarkSkip = NULL, extLst = NULL, ...)
}
\arguments{
\item{tickLblSkip}{tickLblSkip}

\item{tickMarkSkip}{tickMarkSkip}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
SeriesAxis()
}
}
