% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriesFactory.R
\name{SeriesFactory}
\alias{SeriesFactory}
\title{SeriesFactory}
\usage{
SeriesFactory(
  values,
  xvalues = NULL,
  zvalues = NULL,
  title = NULL,
  title_from_data = FALSE,
  ...
)
}
\arguments{
\item{values}{values}

\item{xvalues}{xvalues}

\item{zvalues}{zvalues}

\item{title}{title}

\item{title_from_data}{title_from_data}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Convenience Factory for creating chart data series.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.series_factory.html}
}
\examples{
\dontrun{
SeriesFactory(values = "sheetname!A1:E4")
}
}
