% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriesLabel.R
\name{SeriesLabel}
\alias{SeriesLabel}
\title{SeriesLabel}
\usage{
SeriesLabel(strRef = NULL, v = NULL, ...)
}
\arguments{
\item{strRef}{strRef}

\item{v}{v}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.series.html}
}
\examples{
\dontrun{
SeriesLabel()
}
}
