% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\name{Shape}
\alias{Shape}
\title{Shape}
\usage{
Shape(
  macro = NULL,
  textlink = NULL,
  fPublished = NULL,
  fLocksText = NULL,
  nvSpPr = NULL,
  spPr = NULL,
  style = NULL,
  txBody = NULL,
  ...
)
}
\arguments{
\item{macro}{macro}

\item{textlink}{textlink}

\item{fPublished}{fPublished}

\item{fLocksText}{fLocksText}

\item{nvSpPr}{nvSpPr}

\item{spPr}{spPr}

\item{style}{style}

\item{txBody}{txBody}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.connector.html}
}
\examples{
\dontrun{
Shape(spPr = GraphicalProperties())
}
}
