% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape3D.R
\name{Shape3D}
\alias{Shape3D}
\title{Shape3D}
\usage{
Shape3D(
  z = NULL,
  extrusionH = NULL,
  contourW = NULL,
  prstMaterial = NULL,
  bevelT = NULL,
  bevelB = NULL,
  extrusionClr = NULL,
  contourClr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{z}{z}

\item{extrusionH}{extrusionH}

\item{contourW}{contourW}

\item{prstMaterial}{prstMaterial}

\item{bevelT}{bevelT}

\item{bevelB}{bevelB}

\item{extrusionClr}{extrusionClr}

\item{contourClr}{contourClr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Shape3D()
}

}
