% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeStyle.R
\name{ShapeStyle}
\alias{ShapeStyle}
\title{ShapeStyle}
\usage{
ShapeStyle(lnRef = NULL, fillRef = NULL, effectRef = NULL, fontRef = NULL, ...)
}
\arguments{
\item{lnRef}{lnRef}

\item{fillRef}{fillRef}

\item{effectRef}{effectRef}

\item{fontRef}{fontRef}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
ShapeStyle(
lnRef = StyleMatrixReference(idx = 1L),
fillRef = StyleMatrixReference(idx = 1L),
effectRef = StyleMatrixReference(idx = 1L),
fontRef = FontReference()
)
}

}
