% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetBackgroundPicture.R
\name{SheetBackgroundPicture}
\alias{SheetBackgroundPicture}
\title{SheetBackgroundPicture}
\usage{
SheetBackgroundPicture(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.picture.html}
}
\examples{
\dontrun{
SheetBackgroundPicture()
}

}
