% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetDimension.R
\name{SheetDimension}
\alias{SheetDimension}
\title{SheetDimension}
\usage{
SheetDimension(ref = NULL, ...)
}
\arguments{
\item{ref}{ref}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.dimensions.html}
}
\examples{
\dontrun{
SheetDimension(ref = "ref")
}

}
