% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetFormatProperties.R
\name{SheetFormatProperties}
\alias{SheetFormatProperties}
\title{SheetFormatProperties}
\usage{
SheetFormatProperties(
  baseColWidth = 8L,
  defaultColWidth = NULL,
  defaultRowHeight = 15L,
  customHeight = NULL,
  zeroHeight = NULL,
  thickTop = NULL,
  thickBottom = NULL,
  outlineLevelRow = NULL,
  outlineLevelCol = NULL,
  ...
)
}
\arguments{
\item{baseColWidth}{baseColWidth}

\item{defaultColWidth}{defaultColWidth}

\item{defaultRowHeight}{defaultRowHeight}

\item{customHeight}{customHeight}

\item{zeroHeight}{zeroHeight}

\item{thickTop}{thickTop}

\item{thickBottom}{thickBottom}

\item{outlineLevelRow}{outlineLevelRow}

\item{outlineLevelCol}{outlineLevelCol}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.dimensions.html}
}
\examples{
\dontrun{
SheetFormatProperties()
}

}
