% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetView.R
\name{SheetView}
\alias{SheetView}
\title{SheetView}
\usage{
SheetView(
  windowProtection = NULL,
  showFormulas = NULL,
  showGridLines = NULL,
  showRowColHeaders = NULL,
  showZeros = NULL,
  rightToLeft = NULL,
  tabSelected = NULL,
  showRuler = NULL,
  showOutlineSymbols = NULL,
  defaultGridColor = NULL,
  showWhiteSpace = NULL,
  view = NULL,
  topLeftCell = NULL,
  colorId = NULL,
  zoomScale = NULL,
  zoomScaleNormal = NULL,
  zoomScaleSheetLayoutView = NULL,
  zoomScalePageLayoutView = NULL,
  zoomToFit = NULL,
  workbookViewId = 0L,
  selection = NULL,
  pane = NULL,
  ...
)
}
\arguments{
\item{windowProtection}{windowProtection}

\item{showFormulas}{showFormulas}

\item{showGridLines}{showGridLines}

\item{showRowColHeaders}{showRowColHeaders}

\item{showZeros}{showZeros}

\item{rightToLeft}{rightToLeft}

\item{tabSelected}{tabSelected}

\item{showRuler}{showRuler}

\item{showOutlineSymbols}{showOutlineSymbols}

\item{defaultGridColor}{defaultGridColor}

\item{showWhiteSpace}{showWhiteSpace}

\item{view}{view}

\item{topLeftCell}{topLeftCell}

\item{colorId}{colorId}

\item{zoomScale}{zoomScale}

\item{zoomScaleNormal}{zoomScaleNormal}

\item{zoomScaleSheetLayoutView}{zoomScaleSheetLayoutView}

\item{zoomScalePageLayoutView}{zoomScalePageLayoutView}

\item{zoomToFit}{zoomToFit}

\item{workbookViewId}{workbookViewId}

\item{selection}{selection}

\item{pane}{pane}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Information about the visible portions of this sheet.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.views.html}
}
\examples{
\dontrun{
SheetView()
}

}
