% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetViewList.R
\name{SheetViewList}
\alias{SheetViewList}
\title{SheetViewList}
\usage{
SheetViewList(sheetView = NULL, extLst = NULL, ...)
}
\arguments{
\item{sheetView}{sheetView}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.views.html}
}
\examples{
\dontrun{
SheetViewList()
}

}
