% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_color.R
\name{short_color}
\alias{short_color}
\title{short_color}
\usage{
short_color(color, ...)
}
\arguments{
\item{color}{color}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
format a color to its short size

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.utils.units.html}
}
\examples{
\dontrun{
short_color(color = "color")
}

}
