% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Side.R
\name{Side}
\alias{Side}
\title{Side}
\usage{
Side(style = NULL, color = NULL, border_style = NULL, ...)
}
\arguments{
\item{style}{style}

\item{color}{color}

\item{border_style}{border_style}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Border options for use in styles.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.borders.html}
}
\details{
Caution: if you do not specify a border_style, other attributes will
have no effect !
}
\section{Caution}{
 if you do not specify a border_style, other attributes will:
have no effect !
}

\examples{
\dontrun{
Side()
}

}
