% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTag.R
\name{SmartTag}
\alias{SmartTag}
\title{SmartTag}
\usage{
SmartTag(namespaceUri = NULL, name = NULL, url = NULL, ...)
}
\arguments{
\item{namespaceUri}{namespaceUri}

\item{name}{name}

\item{url}{url}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.smart_tags.html}
}
\examples{
\dontrun{
SmartTag()
}

}
