% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTagList.R
\name{SmartTagList}
\alias{SmartTagList}
\title{SmartTagList}
\usage{
SmartTagList(smartTagType = list(), ...)
}
\arguments{
\item{smartTagType}{smartTagType}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.smart_tags.html}
}
\examples{
\dontrun{
SmartTagList()
}

}
