% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTagProperties.R
\name{SmartTagProperties}
\alias{SmartTagProperties}
\title{SmartTagProperties}
\usage{
SmartTagProperties(embed = NULL, show = NULL, ...)
}
\arguments{
\item{embed}{embed}

\item{show}{show}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.smart_tags.html}
}
\examples{
\dontrun{
SmartTagProperties()
}

}
