% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTags.R
\name{SmartTags}
\alias{SmartTags}
\title{SmartTags}
\usage{
SmartTags(cellSmartTags = list(), ...)
}
\arguments{
\item{cellSmartTags}{cellSmartTags}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.smart_tag.html}
}
\examples{
\dontrun{
SmartTags()
}

}
