% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftEdgesEffect.R
\name{SoftEdgesEffect}
\alias{SoftEdgesEffect}
\title{SoftEdgesEffect}
\usage{
SoftEdgesEffect(rad = NULL, ...)
}
\arguments{
\item{rad}{rad}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
SoftEdgesEffect(rad = 1L)
}

}
