% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortCondition.R
\name{SortCondition}
\alias{SortCondition}
\title{SortCondition}
\usage{
SortCondition(
  ref = NULL,
  descending = NULL,
  sortBy = NULL,
  customList = NULL,
  dxfId = NULL,
  iconSet = NULL,
  iconId = NULL,
  ...
)
}
\arguments{
\item{ref}{ref}

\item{descending}{descending}

\item{sortBy}{sortBy}

\item{customList}{customList}

\item{dxfId}{dxfId}

\item{iconSet}{iconSet}

\item{iconId}{iconId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
SortCondition()
}

}
