% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortState.R
\name{SortState}
\alias{SortState}
\title{SortState}
\usage{
SortState(
  columnSort = NULL,
  caseSensitive = NULL,
  sortMethod = NULL,
  ref = NULL,
  sortCondition = list(),
  extLst = NULL,
  ...
)
}
\arguments{
\item{columnSort}{columnSort}

\item{caseSensitive}{caseSensitive}

\item{sortMethod}{sortMethod}

\item{ref}{ref}

\item{sortCondition}{sortCondition}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.filters.html}
}
\examples{
\dontrun{
SortState()
}

}
