% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spacing.R
\name{Spacing}
\alias{Spacing}
\title{Spacing}
\usage{
Spacing(spcPct = NULL, spcPts = NULL, ...)
}
\arguments{
\item{spcPct}{spcPct}

\item{spcPts}{spcPts}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
Spacing()
}

}
