% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereCoords.R
\name{SphereCoords}
\alias{SphereCoords}
\title{SphereCoords}
\usage{
SphereCoords(lat = NULL, lon = NULL, rev = NULL, ...)
}
\arguments{
\item{lat}{lat}

\item{lon}{lon}

\item{rev}{rev}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
SphereCoords(lat = 1L, lon = 1L, rev = 1L)
}

}
