% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadsheetDrawing.R
\name{SpreadsheetDrawing}
\alias{SpreadsheetDrawing}
\title{SpreadsheetDrawing}
\usage{
SpreadsheetDrawing(
  twoCellAnchor = list(),
  oneCellAnchor = list(),
  absoluteAnchor = list(),
  ...
)
}
\arguments{
\item{twoCellAnchor}{twoCellAnchor}

\item{oneCellAnchor}{oneCellAnchor}

\item{absoluteAnchor}{absoluteAnchor}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
SpreadsheetDrawing()
}

}
