% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockChart.R
\name{StockChart}
\alias{StockChart}
\title{StockChart}
\usage{
StockChart(
  ser = list(),
  dLbls = NULL,
  dropLines = NULL,
  hiLowLines = NULL,
  upDownBars = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{ser}{ser}

\item{dLbls}{dLbls}

\item{dropLines}{dropLines}

\item{hiLowLines}{hiLowLines}

\item{upDownBars}{upDownBars}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.stock_chart.html}
}
\examples{
\dontrun{
StockChart()
}

}
