% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stop.R
\name{Stop}
\alias{Stop}
\title{Stop}
\usage{
Stop(color, position, ...)
}
\arguments{
\item{color}{color}

\item{position}{position}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.fills.html}
}
\examples{
\dontrun{
Stop(color = Color(), position = 1L)
}

}
