% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StopList.R
\name{StopList}
\alias{StopList}
\title{StopList}
\usage{
StopList(name = NULL, ...)
}
\arguments{
\item{name}{name}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
A sequence (list or list) that may only contain objects of the declared

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.fills.html}
}
\details{
type
}
\examples{
\dontrun{
StopList()
}

}
