% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrData.R
\name{StrData}
\alias{StrData}
\title{StrData}
\usage{
StrData(ptCount = NULL, pt = list(), extLst = NULL, ...)
}
\arguments{
\item{ptCount}{ptCount}

\item{pt}{pt}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
StrData()
}

}
