% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StretchInfoProperties.R
\name{StretchInfoProperties}
\alias{StretchInfoProperties}
\title{StretchInfoProperties}
\usage{
StretchInfoProperties(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.fill.html}
}
\examples{
\dontrun{
StretchInfoProperties()
}

}
