% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrRef.R
\name{StrRef}
\alias{StrRef}
\title{StrRef}
\usage{
StrRef(f = NULL, strCache = NULL, extLst = NULL, ...)
}
\arguments{
\item{f}{f}

\item{strCache}{strCache}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
StrRef()
}

}
