% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrVal.R
\name{StrVal}
\alias{StrVal}
\title{StrVal}
\usage{
StrVal(idx = 0L, v = NULL, ...)
}
\arguments{
\item{idx}{idx}

\item{v}{v}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.data_source.html}
}
\examples{
\dontrun{
StrVal()
}

}
