% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleArray.R
\name{StyleArray}
\alias{StyleArray}
\title{StyleArray}
\usage{
StyleArray(args = c(0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L), ...)
}
\arguments{
\item{args}{args}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Simplified named list with an array

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.cell_style.html}
}
\examples{
\dontrun{
StyleArray()
}

}
