% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stylesheet.R
\name{Stylesheet}
\alias{Stylesheet}
\title{Stylesheet}
\usage{
Stylesheet(
  numFmts = NULL,
  fonts = list(),
  fills = list(),
  borders = list(),
  cellStyleXfs = NULL,
  cellXfs = NULL,
  cellStyles = NULL,
  dxfs = list(),
  tableStyles = NULL,
  colors = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{numFmts}{numFmts}

\item{fonts}{fonts}

\item{fills}{fills}

\item{borders}{borders}

\item{cellStyleXfs}{cellStyleXfs}

\item{cellXfs}{cellXfs}

\item{cellStyles}{cellStyles}

\item{dxfs}{dxfs}

\item{tableStyles}{tableStyles}

\item{colors}{colors}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.stylesheet.html}
}
\examples{
\dontrun{
Stylesheet()
}

}
