% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfaceChart3D.R
\name{SurfaceChart3D}
\alias{SurfaceChart3D}
\title{SurfaceChart3D}
\usage{
SurfaceChart3D(wireframe = NULL, ser = list(), bandFmts = NULL, ...)
}
\arguments{
\item{wireframe}{wireframe}

\item{ser}{ser}

\item{bandFmts}{bandFmts}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.surface_chart.html}
}
\examples{
\dontrun{
SurfaceChart3D()
}

}
