% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableColumn.R
\name{TableColumn}
\alias{TableColumn}
\title{TableColumn}
\usage{
TableColumn(
  id = NULL,
  uniqueName = NULL,
  name = NULL,
  totalsRowFunction = NULL,
  totalsRowLabel = NULL,
  queryTableFieldId = NULL,
  headerRowDxfId = NULL,
  dataDxfId = NULL,
  totalsRowDxfId = NULL,
  headerRowCellStyle = NULL,
  dataCellStyle = NULL,
  totalsRowCellStyle = NULL,
  calculatedColumnFormula = NULL,
  totalsRowFormula = NULL,
  xmlColumnPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{id}{id}

\item{uniqueName}{uniqueName}

\item{name}{name}

\item{totalsRowFunction}{totalsRowFunction}

\item{totalsRowLabel}{totalsRowLabel}

\item{queryTableFieldId}{queryTableFieldId}

\item{headerRowDxfId}{headerRowDxfId}

\item{dataDxfId}{dataDxfId}

\item{totalsRowDxfId}{totalsRowDxfId}

\item{headerRowCellStyle}{headerRowCellStyle}

\item{dataCellStyle}{dataCellStyle}

\item{totalsRowCellStyle}{totalsRowCellStyle}

\item{calculatedColumnFormula}{calculatedColumnFormula}

\item{totalsRowFormula}{totalsRowFormula}

\item{xmlColumnPr}{xmlColumnPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
TableColumn(id = 1L, name = "name")
}

}
