% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFormula.R
\name{TableFormula}
\alias{TableFormula}
\title{TableFormula}
\usage{
TableFormula(array = NULL, attr_text = NULL, ...)
}
\arguments{
\item{array}{array}

\item{attr_text}{attr_text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
TableFormula()
}

}
