% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableNameDescriptor.R
\name{TableNameDescriptor}
\alias{TableNameDescriptor}
\title{TableNameDescriptor}
\usage{
TableNameDescriptor(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
TableNameDescriptor()
}

}
