% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablePartList.R
\name{TablePartList}
\alias{TablePartList}
\title{TablePartList}
\usage{
TablePartList(count = NULL, tablePart = list(), ...)
}
\arguments{
\item{count}{count}

\item{tablePart}{tablePart}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
TablePartList()
}

}
