% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableStyle.R
\name{TableStyle}
\alias{TableStyle}
\title{TableStyle}
\usage{
TableStyle(
  name = NULL,
  pivot = NULL,
  table = NULL,
  count = NULL,
  tableStyleElement = list(),
  ...
)
}
\arguments{
\item{name}{name}

\item{pivot}{pivot}

\item{table}{table}

\item{count}{count}

\item{tableStyleElement}{tableStyleElement}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.table.html}
}
\examples{
\dontrun{
TableStyle(name = "name")
}

}
