% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableStyleElement.R
\name{TableStyleElement}
\alias{TableStyleElement}
\title{TableStyleElement}
\usage{
TableStyleElement(type = NULL, size = NULL, dxfId = NULL, ...)
}
\arguments{
\item{type}{type}

\item{size}{size}

\item{dxfId}{dxfId}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.table.html}
}
\examples{
\dontrun{
TableStyleElement(type = "secondSubtotalRow")
}

}
