% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableStyleInfo.R
\name{TableStyleInfo}
\alias{TableStyleInfo}
\title{TableStyleInfo}
\usage{
TableStyleInfo(
  name = NULL,
  showFirstColumn = NULL,
  showLastColumn = NULL,
  showRowStripes = NULL,
  showColumnStripes = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{showFirstColumn}{showFirstColumn}

\item{showLastColumn}{showLastColumn}

\item{showRowStripes}{showRowStripes}

\item{showColumnStripes}{showColumnStripes}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
TableStyleInfo()
}

}
