% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableStyleList.R
\name{TableStyleList}
\alias{TableStyleList}
\title{TableStyleList}
\usage{
TableStyleList(
  count = NULL,
  defaultTableStyle = "TableStyleMedium9",
  defaultPivotStyle = "PivotStyleLight16",
  tableStyle = list(),
  ...
)
}
\arguments{
\item{count}{count}

\item{defaultTableStyle}{defaultTableStyle}

\item{defaultPivotStyle}{defaultPivotStyle}

\item{tableStyle}{tableStyle}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.table.html}
}
\examples{
\dontrun{
TableStyleList()
}

}
