% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAxis.R
\name{TextAxis}
\alias{TextAxis}
\title{TextAxis}
\usage{
TextAxis(
  auto = NULL,
  lblAlgn = NULL,
  lblOffset = 100L,
  tickLblSkip = NULL,
  tickMarkSkip = NULL,
  noMultiLvlLbl = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{auto}{auto}

\item{lblAlgn}{lblAlgn}

\item{lblOffset}{lblOffset}

\item{tickLblSkip}{tickLblSkip}

\item{tickMarkSkip}{tickMarkSkip}

\item{noMultiLvlLbl}{noMultiLvlLbl}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.axis.html}
}
\examples{
\dontrun{
TextAxis()
}

}
