% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextBlock.R
\name{TextBlock}
\alias{TextBlock}
\title{TextBlock}
\usage{
TextBlock(font, text, ...)
}
\arguments{
\item{font}{font}

\item{text}{text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Represents text string in a specific format

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.cell.rich_text.html}
}
\details{
This class is used as part of constructing a rich text strings.
}
\examples{
\dontrun{
TextBlock(font = InlineFont(), text = "text")
}

}
