% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextField.R
\name{TextField}
\alias{TextField}
\title{TextField}
\usage{
TextField(id = NULL, type = NULL, rPr = NULL, pPr = NULL, t = NULL, ...)
}
\arguments{
\item{id}{id}

\item{type}{type}

\item{rPr}{rPr}

\item{pPr}{pPr}

\item{t}{t}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
TextField(id = "id")
}

}
