% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextNormalAutofit.R
\name{TextNormalAutofit}
\alias{TextNormalAutofit}
\title{TextNormalAutofit}
\usage{
TextNormalAutofit(fontScale = NULL, lnSpcReduction = NULL, ...)
}
\arguments{
\item{fontScale}{fontScale}

\item{lnSpcReduction}{lnSpcReduction}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.text.html}
}
\examples{
\dontrun{
TextNormalAutofit(fontScale = 1L, lnSpcReduction = 1L)
}

}
