% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TintEffect.R
\name{TintEffect}
\alias{TintEffect}
\title{TintEffect}
\usage{
TintEffect(hue = 0L, amt = 0L, ...)
}
\arguments{
\item{hue}{hue}

\item{amt}{amt}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.effect.html}
}
\examples{
\dontrun{
TintEffect()
}

}
