% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Title.R
\name{Title}
\alias{Title}
\title{Title}
\usage{
Title(
  tx = NULL,
  layout = NULL,
  overlay = NULL,
  spPr = NULL,
  txPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{tx}{tx}

\item{layout}{layout}

\item{overlay}{overlay}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.title.html}
}
\examples{
\dontrun{
Title()
}

}
