% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_maker.R
\name{title_maker}
\alias{title_maker}
\title{title_maker}
\usage{
title_maker(text, ...)
}
\arguments{
\item{text}{text}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.title.html}
}
\examples{
\dontrun{
title_maker(text = "text")
}

}
