% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TitleDescriptor.R
\name{TitleDescriptor}
\alias{TitleDescriptor}
\title{TitleDescriptor}
\usage{
TitleDescriptor(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.title.html}
}
\examples{
\dontrun{
TitleDescriptor()
}

}
