% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trendline.R
\name{Trendline}
\alias{Trendline}
\title{Trendline}
\usage{
Trendline(
  name = NULL,
  spPr = NULL,
  trendlineType = "linear",
  order = NULL,
  period = NULL,
  forward = NULL,
  backward = NULL,
  intercept = NULL,
  dispRSqr = NULL,
  dispEq = NULL,
  trendlineLbl = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{spPr}{spPr}

\item{trendlineType}{trendlineType}

\item{order}{order}

\item{period}{period}

\item{forward}{forward}

\item{backward}{backward}

\item{intercept}{intercept}

\item{dispRSqr}{dispRSqr}

\item{dispEq}{dispEq}

\item{trendlineLbl}{trendlineLbl}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.trendline.html}
}
\examples{
\dontrun{
Trendline()
}

}
