% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendlineLabel.R
\name{TrendlineLabel}
\alias{TrendlineLabel}
\title{TrendlineLabel}
\usage{
TrendlineLabel(
  layout = NULL,
  tx = NULL,
  numFmt = NULL,
  spPr = NULL,
  txPr = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{layout}{layout}

\item{tx}{tx}

\item{numFmt}{numFmt}

\item{spPr}{spPr}

\item{txPr}{txPr}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.trendline.html}
}
\examples{
\dontrun{
TrendlineLabel()
}

}
