% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoCellAnchor.R
\name{TwoCellAnchor}
\alias{TwoCellAnchor}
\title{TwoCellAnchor}
\usage{
TwoCellAnchor(editAs = NULL, from = NULL, to = NULL, ...)
}
\arguments{
\item{editAs}{editAs}

\item{from}{_from}

\item{to}{to}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Note that the argument `_from` has been transformed to `from`.

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.spreadsheet_drawing.html}
}
\examples{
\dontrun{
TwoCellAnchor()
}

}
