% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpDownBars.R
\name{UpDownBars}
\alias{UpDownBars}
\title{UpDownBars}
\usage{
UpDownBars(gapWidth = 150L, upBars = NULL, downBars = NULL, extLst = NULL, ...)
}
\arguments{
\item{gapWidth}{gapWidth}

\item{upBars}{upBars}

\item{downBars}{downBars}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.updown_bars.html}
}
\examples{
\dontrun{
UpDownBars()
}

}
