% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueDescriptor.R
\name{ValueDescriptor}
\alias{ValueDescriptor}
\title{ValueDescriptor}
\usage{
ValueDescriptor(...)
}
\arguments{
\item{...}{Additional arguments, i.e. kwargs}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.formatting.rule.html}

Expected type depends upon type attribute of parent :-(

Most values should be numeric BUT they can also be cell references
}
\examples{
\dontrun{
ValueDescriptor()
}

}
