% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vector3D.R
\name{Vector3D}
\alias{Vector3D}
\title{Vector3D}
\usage{
Vector3D(dx = NULL, dy = NULL, dz = NULL, ...)
}
\arguments{
\item{dx}{dx}

\item{dy}{dy}

\item{dz}{dz}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.geometry.html}
}
\examples{
\dontrun{
Vector3D(dx = 1L, dy = 1L, dz = 1L)
}

}
