% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebPublishing.R
\name{WebPublishing}
\alias{WebPublishing}
\title{WebPublishing}
\usage{
WebPublishing(
  css = NULL,
  thicket = NULL,
  longFileNames = NULL,
  vml = NULL,
  allowPng = NULL,
  targetScreenSize = "800x600",
  dpi = NULL,
  codePage = NULL,
  characterSet = NULL,
  ...
)
}
\arguments{
\item{css}{css}

\item{thicket}{thicket}

\item{longFileNames}{longFileNames}

\item{vml}{vml}

\item{allowPng}{allowPng}

\item{targetScreenSize}{targetScreenSize}

\item{dpi}{dpi}

\item{codePage}{codePage}

\item{characterSet}{characterSet}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.web.html}
}
\examples{
\dontrun{
WebPublishing()
}

}
