% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebPublishItem.R
\name{WebPublishItem}
\alias{WebPublishItem}
\title{WebPublishItem}
\usage{
WebPublishItem(
  id = NULL,
  divId = NULL,
  sourceType = NULL,
  sourceRef = NULL,
  sourceObject = NULL,
  destinationFile = NULL,
  title = NULL,
  autoRepublish = NULL,
  ...
)
}
\arguments{
\item{id}{id}

\item{divId}{divId}

\item{sourceType}{sourceType}

\item{sourceRef}{sourceRef}

\item{sourceObject}{sourceObject}

\item{destinationFile}{destinationFile}

\item{title}{title}

\item{autoRepublish}{autoRepublish}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chartsheet.publish.html}
}
\examples{
\dontrun{
WebPublishItem(
id = 1L,
divId = "divId",
sourceType = "chart",
sourceRef = "sourceRef",
destinationFile = "destinationFile"
)
}

}
