% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebPublishObject.R
\name{WebPublishObject}
\alias{WebPublishObject}
\title{WebPublishObject}
\usage{
WebPublishObject(
  id = NULL,
  divId = NULL,
  sourceObject = NULL,
  destinationFile = NULL,
  title = NULL,
  autoRepublish = NULL,
  ...
)
}
\arguments{
\item{id}{id}

\item{divId}{divId}

\item{sourceObject}{sourceObject}

\item{destinationFile}{destinationFile}

\item{title}{title}

\item{autoRepublish}{autoRepublish}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.web.html}
}
\examples{
\dontrun{
WebPublishObject(
id = 1L,
divId = "divId",
destinationFile = "destinationFile"
)
}

}
