% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebPublishObjectList.R
\name{WebPublishObjectList}
\alias{WebPublishObjectList}
\title{WebPublishObjectList}
\usage{
WebPublishObjectList(count = NULL, webPublishObject = list(), ...)
}
\arguments{
\item{count}{count}

\item{webPublishObject}{webPublishObject}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.web.html}
}
\examples{
\dontrun{
WebPublishObjectList()
}

}
