% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workbook.R
\name{Workbook}
\alias{Workbook}
\title{Workbook}
\usage{
Workbook(write_only = FALSE, iso_dates = FALSE, ...)
}
\arguments{
\item{write_only}{write_only}

\item{iso_dates}{iso_dates}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.workbook.html#openpyxl.workbook.workbook.Workbook}
}
\examples{
\dontrun{
Workbook()
}

}
