% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkbookParser.R
\name{WorkbookParser}
\alias{WorkbookParser}
\title{WorkbookParser}
\usage{
WorkbookParser(archive, workbook_part_name, keep_links = TRUE, ...)
}
\arguments{
\item{archive}{archive}

\item{workbook_part_name}{workbook_part_name}

\item{keep_links}{keep_links}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.reader.workbook.html}
}
\examples{
\dontrun{
WorkbookParser(
archive = "archive",
workbook_part_name = "workbook_part_name"
)
}

}
