% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkbookProperties.R
\name{WorkbookProperties}
\alias{WorkbookProperties}
\title{WorkbookProperties}
\usage{
WorkbookProperties(
  date1904 = NULL,
  dateCompatibility = NULL,
  showObjects = NULL,
  showBorderUnselectedTables = NULL,
  filterPrivacy = NULL,
  promptedSolutions = NULL,
  showInkAnnotation = NULL,
  backupFile = NULL,
  saveExternalLinkValues = NULL,
  updateLinks = NULL,
  codeName = NULL,
  hidePivotFieldList = NULL,
  showPivotChartFilter = NULL,
  allowRefreshQuery = NULL,
  publishItems = NULL,
  checkCompatibility = NULL,
  autoCompressPictures = NULL,
  refreshAllConnections = NULL,
  defaultThemeVersion = NULL,
  ...
)
}
\arguments{
\item{date1904}{date1904}

\item{dateCompatibility}{dateCompatibility}

\item{showObjects}{showObjects}

\item{showBorderUnselectedTables}{showBorderUnselectedTables}

\item{filterPrivacy}{filterPrivacy}

\item{promptedSolutions}{promptedSolutions}

\item{showInkAnnotation}{showInkAnnotation}

\item{backupFile}{backupFile}

\item{saveExternalLinkValues}{saveExternalLinkValues}

\item{updateLinks}{updateLinks}

\item{codeName}{codeName}

\item{hidePivotFieldList}{hidePivotFieldList}

\item{showPivotChartFilter}{showPivotChartFilter}

\item{allowRefreshQuery}{allowRefreshQuery}

\item{publishItems}{publishItems}

\item{checkCompatibility}{checkCompatibility}

\item{autoCompressPictures}{autoCompressPictures}

\item{refreshAllConnections}{refreshAllConnections}

\item{defaultThemeVersion}{defaultThemeVersion}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.properties.html}
}
\examples{
\dontrun{
WorkbookProperties()
}

}
