% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkbookProtection.R
\name{WorkbookProtection}
\alias{WorkbookProtection}
\title{WorkbookProtection}
\usage{
WorkbookProtection(
  workbookPassword = NULL,
  workbookPasswordCharacterSet = NULL,
  revisionsPassword = NULL,
  revisionsPasswordCharacterSet = NULL,
  lockStructure = NULL,
  lockWindows = NULL,
  lockRevision = NULL,
  revisionsAlgorithmName = NULL,
  revisionsHashValue = NULL,
  revisionsSaltValue = NULL,
  revisionsSpinCount = NULL,
  workbookAlgorithmName = NULL,
  workbookHashValue = NULL,
  workbookSaltValue = NULL,
  workbookSpinCount = NULL,
  ...
)
}
\arguments{
\item{workbookPassword}{workbookPassword}

\item{workbookPasswordCharacterSet}{workbookPasswordCharacterSet}

\item{revisionsPassword}{revisionsPassword}

\item{revisionsPasswordCharacterSet}{revisionsPasswordCharacterSet}

\item{lockStructure}{lockStructure}

\item{lockWindows}{lockWindows}

\item{lockRevision}{lockRevision}

\item{revisionsAlgorithmName}{revisionsAlgorithmName}

\item{revisionsHashValue}{revisionsHashValue}

\item{revisionsSaltValue}{revisionsSaltValue}

\item{revisionsSpinCount}{revisionsSpinCount}

\item{workbookAlgorithmName}{workbookAlgorithmName}

\item{workbookHashValue}{workbookHashValue}

\item{workbookSaltValue}{workbookSaltValue}

\item{workbookSpinCount}{workbookSpinCount}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.workbook.protection.html}
}
\examples{
\dontrun{
WorkbookProtection()
}

}
