% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorksheetCopy.R
\name{WorksheetCopy}
\alias{WorksheetCopy}
\title{WorksheetCopy}
\usage{
WorksheetCopy(source_worksheet, target_worksheet, ...)
}
\arguments{
\item{source_worksheet}{source_worksheet}

\item{target_worksheet}{target_worksheet}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Copy the values, styles, dimensions, merged cells, margins, and

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.copier.html}
}
\details{
print/page setup from one worksheet to another within the same
workbook.
}
\examples{
\dontrun{
wb <- Workbook()
ws1 <- wb$active
ws2 <- wb$create_sheet("Mysheet")
WorksheetCopy(source_worksheet = ws1, target_worksheet = ws2)
}

}
