% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorksheetProperties.R
\name{WorksheetProperties}
\alias{WorksheetProperties}
\title{WorksheetProperties}
\usage{
WorksheetProperties(
  codeName = NULL,
  enableFormatConditionsCalculation = NULL,
  filterMode = NULL,
  published = NULL,
  syncHorizontal = NULL,
  syncRef = NULL,
  syncVertical = NULL,
  transitionEvaluation = NULL,
  transitionEntry = NULL,
  tabColor = NULL,
  outlinePr = NULL,
  pageSetUpPr = NULL,
  ...
)
}
\arguments{
\item{codeName}{codeName}

\item{enableFormatConditionsCalculation}{enableFormatConditionsCalculation}

\item{filterMode}{filterMode}

\item{published}{published}

\item{syncHorizontal}{syncHorizontal}

\item{syncRef}{syncRef}

\item{syncVertical}{syncVertical}

\item{transitionEvaluation}{transitionEvaluation}

\item{transitionEntry}{transitionEntry}

\item{tabColor}{tabColor}

\item{outlinePr}{outlinePr}

\item{pageSetUpPr}{pageSetUpPr}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.properties.html}
}
\examples{
\dontrun{
WorksheetProperties()
}

}
