% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stylesheet.R
\name{write_stylesheet}
\alias{write_stylesheet}
\title{write_stylesheet}
\usage{
write_stylesheet(wb, ...)
}
\arguments{
\item{wb}{wb}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.styles.stylesheet.html}
}
\examples{
\dontrun{
wb <- Workbook()
write_stylesheet(wb = wb)
}

}
