% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Export to xlsx}
\arguments{
\item{x}{data frame or named list of data frames that will be sheets in the xlsx}

\item{path}{A function call using the magrittr semantics.}

\item{col_names}{write column names at the top of the file?}

\item{format_headers}{make the col_names in the xlsx centered and bold}

\item{use_zip64}{use zip64 to enable support for 4GB+ xlsx files. Not all platforms can read this.}
}
\value{
`write_xlsx()` returns the input invisibly.
}
\description{
See \code{writexl::\link[writexl:write_xlsx]{write_xlsx}} for details.
}
\keyword{internal}
