% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_axis.R
\name{x_axis}
\alias{x_axis}
\title{x_axis}
\usage{
x_axis(py_obj, ...)
}
\arguments{
\item{py_obj}{py_obj}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Values must be of type <class 'openpyxl.chart.axis.TextAxis'>
}
\examples{
\dontrun{
BarChart(type = "col") |>
  x_axis(title = "X axis title")
}

}
