% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XDRPoint2D.R
\name{XDRPoint2D}
\alias{XDRPoint2D}
\title{XDRPoint2D}
\usage{
XDRPoint2D(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.drawing.xdr.html}
}
\examples{
\dontrun{
XDRPoint2D(x = 1L, y = 1L)
}

}
