% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLColumnProps.R
\name{XMLColumnProps}
\alias{XMLColumnProps}
\title{XMLColumnProps}
\usage{
XMLColumnProps(
  mapId = NULL,
  xpath = NULL,
  denormalized = NULL,
  xmlDataType = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{mapId}{mapId}

\item{xpath}{xpath}

\item{denormalized}{denormalized}

\item{xmlDataType}{xmlDataType}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.worksheet.table.html}
}
\examples{
\dontrun{
XMLColumnProps(mapId = 1L, xpath = "xpath", xmlDataType = "xmlDataType")
}

}
