% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYSeries.R
\name{XYSeries}
\alias{XYSeries}
\title{XYSeries}
\usage{
XYSeries(
  idx = 0L,
  order = 0L,
  tx = NULL,
  spPr = NULL,
  pictureOptions = NULL,
  dPt = list(),
  dLbls = NULL,
  trendline = NULL,
  errBars = NULL,
  cat = NULL,
  val = NULL,
  invertIfNegative = NULL,
  shape = NULL,
  xVal = NULL,
  yVal = NULL,
  bubbleSize = NULL,
  bubble3D = NULL,
  marker = NULL,
  smooth = NULL,
  explosion = NULL,
  extLst = NULL,
  ...
)
}
\arguments{
\item{idx}{idx}

\item{order}{order}

\item{tx}{tx}

\item{spPr}{spPr}

\item{pictureOptions}{pictureOptions}

\item{dPt}{dPt}

\item{dLbls}{dLbls}

\item{trendline}{trendline}

\item{errBars}{errBars}

\item{cat}{cat}

\item{val}{val}

\item{invertIfNegative}{invertIfNegative}

\item{shape}{shape}

\item{xVal}{xVal}

\item{yVal}{yVal}

\item{bubbleSize}{bubbleSize}

\item{bubble3D}{bubble3D}

\item{marker}{marker}

\item{smooth}{smooth}

\item{explosion}{explosion}

\item{extLst}{extLst}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Dedicated series for charts that have x and y series

\url{https://openpyxl.readthedocs.io/en/stable/api/openpyxl.chart.series.html}
}
\examples{
\dontrun{
XYSeries()
}

}
