% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_axis.R
\name{y_axis}
\alias{y_axis}
\title{y_axis}
\usage{
y_axis(py_obj, ...)
}
\arguments{
\item{py_obj}{py_obj}

\item{...}{Additional arguments, i.e. kwargs.}
}
\value{
An openpyxl Python object.
}
\description{
Values must be of type <class 'openpyxl.chart.axis.NumericAxis'>
}
\examples{
\dontrun{
BarChart(type = "col") |>
  y_axis(title = "Y axis title")
}

}
